% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{airquality}
\alias{airquality}
\title{Air quality data, from the 'datasets' library.}
\format{
Contains six variables, as follows:
\describe{
  \item{ozone}{Ozone (ppb).}
  \item{solar}{Solar R (lang).}
\item{wind}{Wind (mph).}
\item{temp}{Temperature (degrees F).}
\item{month}{Month (1--12).}
\item{day}{Day of the month (1--31).}
 }
}
\source{
The data were obtained from the library 'datasets'.
}
\usage{
data(airquality)
}
\description{
Daily air quality measurements in New York, May to September 1973. 
 Be aware that some variables have missing values.
}
\examples{
data(airquality)    
summary(airquality) 
}
\references{
Chambers J, Cleveland W, Kleiner B, Tukey P. 1983. Graphical Methods for Data Analysis. Belmont. CA: Wadsworth.
}
\keyword{datasets}
