% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{airquality2}
\alias{airquality2}
\title{Contiene informacion de calidad del aire.}
\format{
Contiene 6 variables:
\describe{
  \item{ozone}{Ozono (ppb).}
  \item{solar}{Solar R (largo).}
\item{wind}{Viento (mph).}
\item{temp}{Temperatura (grados F).}
\item{month}{Mes del anho (1--12).}
\item{day}{Dia del mes (1--31).}
 }
}
\source{
Los datos fueron obtenidos desde la libreria 'datasets'.
}
\usage{
data(airquality)
}
\description{
Calidad del aire diario medido en New York, de Mayo a Septiembre de 1973.
}
\examples{
data(airquality2)    
head(airquality2) 
}
\references{
Chambers J, Cleveland W, Kleiner B, Tukey P. 1983. Graphical Methods for Data Analysis. Belmont. CA: Wadsworth.
}
\keyword{datasets}
