% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{annualppCities}
\alias{annualppCities}
\title{Time series of annual precipitations in cities of Chile.}
\format{
The data frame contains three variables as follows:
\describe{
  \item{city}{Name of city.}
  \item{year}{Year of registry.}
  \item{annual}{Value of the annual precipitation of a given year (mm).}
 }
}
\source{
The data were obtained from \url{https://explorador.cr2.cl/}.
}
\usage{
annualppCities
}
\description{
The data frame has annual precipitation in six cities of 
Chile (i.e., Santiago, Talca, Chillan, Temuco, Valdivia, and Puerto Montt) in 
different years, from 1950 to 2019.
 @usage
data(annualppCities)
}
\examples{
head(annualppCities)    
range(annualppCities$year)
}
\keyword{datasets}
