% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{annualppCities2}
\alias{annualppCities2}
\title{Serie de tiempo de precipitaciones anuales en Chile.}
\format{
La data frame tiene tres variables:
\describe{
  \item{ciudad}{Nombre de la ciudad.}
  \item{anho}{Año de registro.}
  \item{pp.anual}{Valor de la precipitación anual (mm).}
 }
}
\source{
Los datos fueron obtenidos desde \url{https://explorador.cr2.cl/}.
}
\usage{
data(annualppCities2)
}
\description{
Precipitaciones anuales en seis ciudades de Chile (i.e., Santiago, Talca, 
Chillan, Temuco, Valdivia, and Puerto Montt) registradas en distintos años, 
desde 1950 al 2019.
}
\examples{
head(annualppCities2)    
range(annualppCities2$anho)
}
\keyword{datasets}
