% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{anscombe2}
\alias{anscombe2}
\title{Cuarteto de Anscombe}
\format{
Variables se describen a continuacion::
\describe{
\item{X1}{Valores enteros que representan el eje X para las columnas Y1, Y2 e Y3}								
\item{Y1}{Valores flotantes que representan el eje Y para la columna X1}
\item{Y2}{Valores flotantes que representan el eje Y para la columna X1}
\item{Y3}{Valores flotantes que representan el eje Y para la columna X1}
\item{X2}{Valores enteros que representan el eje X para las columnas Y4}
\item{Y4}{Valores flotantes que representan el eje Y para la columna X2}
 }
}
\source{
Datos fueron contribuidos por el Dr Christian Salas-Eljatib (Universidad Mayor, Santiago, Chile).
}
\usage{
data(anscombe2)
}
\description{
Dataset que contiene 4 pares de columnas con la mismos estadisticos descriptivos,
 sin embargo si existe diferencia al representar los puntos mediante un grafico.
}
\examples{
data(anscombe2)    
head(anscombe2) 
}
\references{
Anscombe FJ. 1973. Graphs in statistical analysis. The American Statistician, 27, 17-21. \doi{10.2307/2682899}.
}
\keyword{datasets}
