% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{baitreeline2}
\alias{baitreeline2}
\title{Incremento anual en area basal de cuatro especies arboreas.}
\format{
Contains seven columns, as follows:
\describe{
  \item{clima}{Climate of each location, mediterranean and temperate.}
  \item{sitio}{Name of Location of study (termmas:Termas de Chillan, 
  antillanca:Antillanca area within Puyehue National Park,
              castillo:Cerro Castillo Natural Reserve, farellones:Farellones 
              in Central Chile, 
              pyrenees: Sierra de Cutas area in Spanish Central Pyrenees,
              sierra:Sierra Nevada).}
\item{especie}{name species of study (lenga: Nothofagus pumilio, frangel: 
Kageneckia angustifolia,
               uncinata: Pinus uncinata, sylvestris: Pinus sylvestris).}
\item{tipo.altitud}{Type of elevation. "Treeline", intermediate named as 
"inter", and closed or
      montane forest named as low.}
\item{arbol}{Id for tree.}
\item{bai}{Value of annual basal area increment.}
\item{bai.medio}{Mean of annual basal area increment.}
 }
}
\source{
The data were obtained from the DRYAD repository
 at \doi{10.5061/dryad.ks97h}.
}
\usage{
data(baitreeline2)
}
\description{
Este set de datps contiene 157 observaciones, de los ultimos 10 años
 en 6-8 árboles adultos de cuatro especies 
en un gradiente altitudinal. Las muestras se distribuyeron en cuatro
 localidades o sitios de Chile y dos en España.
}
\examples{
data(baitreeline2)    
head(baitreeline2) 
}
\references{
Piper F, Vinegla B, Linares J, Camarero J, Cavieres L, Fajardo A. 2016. 
Mediterranean and temperate treelines are 
controlled by different environmental drivers. Journal of Ecology
 104: 691-702.
}
\keyword{datasets}
