% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{baiTreelines}
\alias{baiTreelines}
\title{Annual basal area increment for four tree species.}
\format{
Contains seven columns, as follows:
\describe{
  \item{climate}{Climate of each location, mediterranean and temperate.}
  \item{site}{Name of Location of study (termmas:Termas de Chillan , antillanca:Antillanca area within Puyehue National Park,
              castillo:Cerro Castillo Natural Reserve, farellones:Farellones in Central Chile, 
              pyrenees: Sierra de Cutas area in Spanish Central Pyrenees,sierra:Sierra Nevada).}
\item{species}{name species of study (lenga: Nothofagus pumilio, frangel: Kageneckia angustifolia,
               uncinata: Pinus uncinata, sylvestris: Pinus sylvestris).}
\item{elevation}{Type of elevation. "Treeline", intermediate named as "inter", and closed or
      montane forest named as low.}
\item{tree}{Id for tree.}
\item{bai}{Value of annual basal area increment.}
\item{mean.bai}{Mean of annual basal area increment.}
 }
}
\source{
The data were obtained from the DRYAD repository at \doi{10.5061/dryad.ks97h}.
}
\usage{
data(baiTreelines)
}
\description{
The dataset contains 157 observations of the last ten years in 6-8 adult trees of different species at three
 elevations of altitudinal gradients sampled in four locations in Chile and two in Spain.
}
\examples{
data(baiTreelines)    
head(baiTreelines) 
}
\references{
Piper F, Vinegla B, Linares J, Camarero J, Cavieres L, Fajardo A. 2016. Mediterranean and temperate treelines are 
controlled by different environmental drivers. Journal of of Ecology. 104: 691-702.
}
\keyword{datasets}
