% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{bears2}
\alias{bears2}
\title{Age and physical measurement data for wild bears.}
\format{
Contiene variables de nivel individual, como se describen a continuacion:
\describe{
  \item{id}{Identificador de oso}
\item{age}{edad en meses}
\item{month}{diametro a la altura del pecho, en cm}
\item{sex}{1 = hombre, 2 = mujer}
\item{headL}{longitud de la cabeza, en cm}
\item{headW}{ancho de la cabeza, en cm}
\item{neckG}{circunferencia del cuello, en cm}
\item{length}{longitud del cuerpo, en cm}
\item{chestG}{circunferencia del pecho, en cm}
\item{weight}{peso corporal, en kg}
\item{obs}{numero de observación para el oso}
\item{name}{nombre dado al oso}
 }
}
\source{
Minitab, Inc. The data description is courtesy of Prof. Timothy Gregoire at
 Yale University (USA).
}
\usage{
data(bears2)
}
\description{
Los osos salvajes fueron anestesiados y sus cuerpos fueron medidos y pesados. Uno de los objetivos del estudio fue hacer una
tabla (o quizas un conjunto de tablas) para las personas interesadas en estimar el peso de un oso basandose en otras medidas. 
Esto se usara porque en el bosque es mas facil medir la longitud de un oso, por ejemplo, que pesarlo.
Observe que faltan valores para algunas de las variables.
}
\examples{
data(bears2)    
head(bears2) 
}
\references{
According to Prof. Gregoire, This data set was supplied by Gary Alt. Entertaining 
 references are in Reader's Digest April, 1979, and Sports Afield September, 1981.
}
\keyword{datasets}
