% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{biomass2}
\alias{biomass2}
\title{Biomasa a nivel de árbol para especies arboreas de Canada.}
\format{
\describe{
  \item{arbol}{Numero del árbol.}
  \item{spp}{Nombre comun de la especie.}
  \item{dap}{Diametro a la altura del pecho, en cm.}
  \item{altura}{Altura total, en m.}
  \item{biomtot}{Biomasa total, en kg.}
  \item{biomfuste}{Biomasa del fuste, en kg.}
  \item{biomramas}{Biomasa de las ramas, en kg.}
  \item{biomfollaje}{Biomasa del follaje, en kg.}
 }
}
\source{
Los datos fueron cedidos por el Prof. Timothy Gregoire de la 
School of Forestry and Environmental Studies de Yale University (New Haven, CT, USA).
}
\usage{
biomass2
}
\description{
Mediciones de componentes de biomasa de árboles muestra de cuatro especies. 
Las especies son Balsam fir (Abies balsamea), Black spruce (Picea mariana),
White birch (Betula papyrifera), y White spruce (Picea glauca).
}
\examples{
head(biomass2)    
table(biomass2$spp)
}
\keyword{datasets}
