% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-carAccidents.r
\docType{data}
\name{carAccidents}
\alias{carAccidents}
\title{Driver status after car accidents in Greece.}
\format{
Contains the factor variables:
\describe{
\item{record}{factor representing the driver status.}
\item{seatBelt}{factor indicating whether the driver wore a setbelt.}
}
}
\source{
R package 'gginference'
}
\usage{
data(carAccidents)
}
\description{
A data frame showing the use of seat belt and the driver status
after a car accident in Greece.
}
\examples{
data(carAccidents)    
head(carAccidents)
table(carAccidents)
}
\keyword{datasets}
