% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDataDatana.r
\docType{data}
\name{carbohtrees2}
\alias{carbohtrees2}
\title{Concentración de carbohidratos de especies arbóreas}
\format{
Hay 16 variables disponibles:
\describe{
  \item{clima}{Tipo de clima de cada sitio: mediterraneo o temperado.}
  \item{sitio}{Nombre del sitio de estudio, como sigue: "termas" (Termas de
   Chillán), "antillanca" (sector
  Antillanca dentro del Parque Nacional), "castillo" (Reserva Nacional Cerro
   Castillo), "farellones" (Farellones, a 20 Kms de Santiago,
  en Chile), "pyrenees" (Sierra de Cutas area in Spanish Central Pyrenees), 
  "sierra" (Sierra Nevada).}
\item{especie}{name species of study (lenga: Nothofagus pumilio, frangel:
 Kageneckia angustifolia,
               uncinata: Pinus uncinata, sylvestris: Pinus sylvestris).}
\item{tejido}{Type of tissue, new developing twings, stem sapwood and 
branches.}
\item{temporada}{Meauserement season (spring or autumn).}
\item{altitud}{Type of elevation. "Treeline", intermediate named as "mid", 
and closed or
      montane forest named as "low".}
\item{arbol}{Id for tree.}
\item{arb.sitio}{Id site for each location of study.}
\item{cts}{Concentración de carbohidratos solubles totales por unidad
 de masa.}
\item{almidon}{Concentración de almidon por unidad de masa.}
\item{cne}{Concentración de carbohidratos no-estructurales por unidad
 de masa.}
\item{cts.cne}{División entre cts y cne.}
\item{dmade}{It might be 'wood density', but not sure.}
\item{tss.mv}{Value of concentrations soluble carbohydrate per volume unit.}
\item{st.mv}{Value of concentrations starch per volume unit.}
\item{nsc.mv}{Value of concentrations non structural carbohydrates per 
volume unit.}
 }
}
\source{
Los datos fueron obtenidos desde el repositorio
 DRYAD en \doi{10.5061/dryad.ks97h}.
}
\usage{
data(carbohtrees2)
}
\description{
Los datos contienen 863 observaciones, sobre carbohidratos totales solubles, 
almidon, y carbohidratos no-estructurales 
 por unidad de masa y por unidad de volumen, en tres tejidos obtenidos al
  comienzo del verano y al comienzo del otoño. Lo
 anterior fue medido entre 6-8 árboles adultos de diferentes especies en un
  gradiente altitudinal, muestreados en
 cuatro sitios en Chile y España.
}
\examples{
data(carbohtrees2)    
head(carbohtrees2) 
}
\references{
Piper F, Vinegla B, Linares J, Camarero J, Cavieres L, Fajardo A. 2016. 
Mediterranean and temperate treelines are controlled by different 
environmental drivers. Journal of Ecology 104:691-702.  
\doi{10.1111/1365-2745.12555}
}
\keyword{datasets}
