% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-caribou.r
\docType{data}
\name{caribou}
\alias{caribou}
\title{Caribou survival}
\format{
Data frame con 91 filas y 3 columnas:
\describe{
\item{herd}{Herd identifier.}
\item{wolf.density}{Wolf density of the herd as  wolf / 100
km².}
\item{alive}{Caribou survival, \code{1} survives, \code{0} don't survive.}
}
}
\usage{
caribou
}
\description{
Caribou survival
}
\examples{
data(caribou)
table(caribou$alive, caribou$herd)
}
\keyword{datasets}
