% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-caribou2.r
\docType{data}
\name{caribou2}
\alias{caribou2}
\title{Sobrevivencia de caribú}
\format{
Data frame con 91 filas y 3 columnas:
\describe{
\item{herd}{Identificador de la manada.}
\item{wolf.density}{Densidad de lobos, en número de lobos / 100
km².}
\item{alive}{Sobrevivencia de un caribú, \code{1} sobrevive, \code{0} no
sobrevive.}
}
}
\usage{
caribou2
}
\description{
Sobrevivencia de caribú
}
\examples{
data(caribou2)
table(caribou2$alive, caribou2$herd)
}
\keyword{datasets}
