% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.R
\name{cdf}
\alias{cdf}
\title{Computes the cumulative distribution for a given random variable}
\usage{
cdf(y = y)
}
\arguments{
\item{y}{a vector of a random variable}
}
\value{
The function returns a dataframe having two columns: the first contains 
the random variable values and the second the cumulative distribution for the variable.
}
\description{
Cumulative distribution for a given random variable.
}
\details{
The cumulative distribution of the random variable is build every 5% of
 the data at hand, that is to say, from 0.05 (i.e., 5%) to 0.95 (i.e., 95%).
}
\examples{
y.var <- rnorm(10)
cdf(y.var)
}
\author{
Christian Salas-Eljatib
}
