% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-chicksw2.r
\docType{data}
\name{chicksw2}
\alias{chicksw2}
\title{Crecimiento de pollos.}
\format{
Contine cuatro variables, como sigue:
\describe{
\item{pollo}{Un identificador único para cada pollo. La numeracion esta
ordenado segun el peso final dentro de cada dieta.}
\item{dieta}{Un factor con cuatro nivels: 1,2,3 y 4 indicando que dieta
recibió el pollo.}
\item{tiempo}{Número de días desde el nacimiento.}
\item{peso}{Peso del pollo (gm).}
}
}
\source{
Los datos fueron obtenidos desde la librería \eqn{alr4}.
}
\usage{
data(chicksw2)
}
\description{
El peso de pollos fueron medidos al momento de nacer y cada dia por medio
hasta el dia 20.
Ellos también fueron medidos el día 21. Hubo cuatro grupos de pollos en
diferentes dietas de proteinas.
}
\examples{
data(chicksw2)    
head(chicksw2) 
}
\references{
Crowder M, Hand D. 1990. Analysis of Repeated Measures. Chapman and Hall
}
\keyword{datasets}
