% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datana-package.R
\docType{package}
\name{datana-package}
\alias{datana}
\alias{datana-package}
\title{datana: Datasets and Functions to Accompany Analisis De Datos Con R}
\description{
Datasets and functions to accompany the book 'Analisis de datos con el programa estadistico R: una introduccion aplicada' by Salas-Eljatib (2021, ISBN: 9789566086109). The package helps carry out data management, exploratory analyses, and model fitting.
}
\author{
\strong{Maintainer}: Christian Salas-Eljatib \email{cseljatib@gmail.com} (\href{https://orcid.org/0000-0002-8468-0829}{ORCID})

Other contributors:
\itemize{
  \item Nicolás Campos (\href{https://orcid.org/0009-0001-9534-9808}{ORCID}) (since 2025) [contributor]
  \item Joaquín Riquelme (up to 2020) [contributor]
  \item Nicolás Pino (up to 2020) [contributor]
}

}
\keyword{internal}
