% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{demograph}
\alias{demograph}
\title{Contains information of demography of species.}
\format{
Contains 15 variables, as follows:
\describe{
  \item{sp}{Name specie.}
  \item{site}{Name of site of study.}
\item{country}{Name of country.}
\item{site.n}{Code of site.}
\item{code}{Code of specie.}
\item{genus}{Genus of specie.}
\item{sps}{Abbreviated name specie.}
\item{family}{Family of specie.}
\item{phyl}{Type of phylogeny.}
\item{l.hab}{Type of leaf habit.}
\item{l.type}{.}
\item{leaf}{Type of leaf.}
\item{growth.l}{Growth at full light (time in years).}
\item{growth.d}{Growth in shade.}
\item{surv.d}{Survival in shade.}
 }
}
\source{
The data were obtained from the DRYAD repository.
}
\usage{
data(demograph)
}
\description{
Dataset contains 61 observations about life histories values for each species and site,
 as obtained from the parameterization carried out in studies that used the model SORTIE
}
\examples{
data(demograph)    
head(demograph) 
}
\references{
- Ameztegui A, Paquette A, Shipley B, Heym M, Messier C, Gravel D. 2016 . Shade tolerance and 
 the functional trait: demography relationship in temperate and boreal forests. Functional Ecology, 
 31: 821-830.
}
\keyword{datasets}
