% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{demograph2}
\alias{demograph2}
\title{Información demográfica de especies arboreas.}
\format{
Contiene 15 variables, como sigue:
\describe{
  \item{especie}{Nombre de la especie.}
  \item{sitio}{Nombre del sitio.}
\item{pais}{Pais.}
\item{sitio.num}{Codigo del sitio.}
\item{codigo}{Codigo de la especie.}
\item{genero}{Genero de la especie.}
\item{spp}{Codigo abreviado de la especie.}
\item{famila}{Familia de la especie.}
\item{filogenia}{Tipo de filogenia: Angiosperma/Gimnosperma}
\item{h.hab}{Tipo de habitat de la hoja: Decidua/Siempreverde.}
\item{h.tipo}{Tipo de hoja.}
\item{hoja}{Combinacion de filogenia y tipo de habitat de la hoja.}
\item{creci.luz}{Growth at full light, in years. Defined as the time needed for
 a sapling with DBH= 1cm to become an adult (DBH=7.5 cm) when light = 100\%.}
\item{creci.sombra}{Growth in shade, in years. Dedined as the time needed for a 
sapling with DBH= 1cm to become an adult (DBH=7.5 cm) when light = 1\%}
\item{sobre.sombra}{Survival in shade, en \%. Defined as the 5-year survivorship of a sapling
 with DBH = 1cm when light = 1\%.}
 }
}
\source{
Los datos se obtuvieron desde el repositorio Dryad \doi{10.5061/dryad.12b0h}.
}
\usage{
data(demograph2)
}
\description{
Contiene 61 observaciones sobre variables de historia de vida segun especie y sitio,
 segun el estilo de los estudios que se usan en el modelo de crecimiento forestal SORTIE.
}
\examples{
data(demograph2)    
head(demograph2) 
}
\references{
- Ameztegui A, Paquette A, Shipley B, Heym M, Messier C, Gravel D. 2016. Shade 
 tolerance and the functional trait: demography relationship in temperate
  and boreal forests. Functional Ecology 31:821-830.
}
\keyword{datasets}
