% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{election}
\alias{election}
\title{Presidential election data of Florida (USA) in 2000.}
\format{
Contains three variables, as follows:
\describe{
  \item{gore}{Vote for Gore.}
  \item{bush}{Vote for Bush.}
\item{buchanan}{Vote for Pat Buchanan.}
 }
}
\source{
The data were obtained from the \eqn{alr4} library.
}
\usage{
data(election)
}
\description{
County-by-county vote for president in Florida in 2000 for Bush, Gore 
and Buchanan.
}
\examples{
data(election)    
head(election) 
}
\references{
Weisberg S. 2014. Applied Linear Regression. 4th edition. Hoboken NJ: Wiley
}
\keyword{datasets}
