% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{eucaleafAll2}
\alias{eucaleafAll2}
\title{Mediciones foliares (todas, n=744) para árboles de Eucalyptus nitens en Tasmania, Australia.}
\format{
Contiene variables a nivel de hoja, como sigue:
\describe{
  \item{tiempo}{factor a dos niveles: Temprano o Tardio}
\item{arbol}{un identificador del árbol muestra}
\item{meristema}{descripcion del meristema}
\item{largo}{largo de la hoja, en mm}
\item{ancho}{ancho de la hoja, en mm}
\item{area}{area foliar, en cm\eqn{^{2}}{^2}}
 }
}
\source{
Aunque la fuente original de estas mediciones proviene de la tesis del Dr. Candy (1999), 
 el archivo de datos fue cortesia del Prof. Timothy Gregoire de
 Yale University (New Haven, CT, USA).
}
\usage{
data(eucaleafAll2)
}
\description{
Mediciones de largo, ancho y area de hojas de Eucalyptus nitens para toda la muestra de Candy (1999).
}
\examples{
data(eucaleafAll2)    
head(eucaleafAll2) 
}
\references{
- Candy SG. 1999. Predictive models for integrated pest management of the
leaf beetle Chrysophtharta bimaculata in Eucalyptus nitens in Tasmania. Doctoral dissertation,
 University of Tasmania, Hobart, Australia.
}
\keyword{datasets}
