% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{eucaplotsampleh2}
\alias{eucaplotsampleh2}
\title{Una parcela de muestreo en una plantación de Eucalyptus globulus, con medición de altura en una submuestra de árboles.}
\format{
Los datos contienen las siguientes columnas:
\describe{
  \item{dap}{Diametro a la altura del pecho, en cm.}
  \item{sanidad}{clase de sanidad (1: buena, 2: media, 3: mala).}
  \item{forma}{clase de forma fustal (1: buena, 2: media, 3: mala).}
  \item{dominancia}{Dominancia (1: superior, 2: intermedio, 3: inferior).}
  \item{atot}{Altura total, en m. Solo medida en una submuestra de árboles.}
 }
}
\source{
Los datos son cedidos por el Dr Christian Salas-Eljatib (Chile).
}
\usage{
data(eucaplotsampleh2)
}
\description{
Datos de una parcela de muestreo, pero las alturas fueron solo medidos en una submuestra de árboles. 
La parcela de muestreo (de 500 m\eqn{^{2}}{^2} de superficie) fue establecidad en una plantación forestal de Eucalyptus globulus cerca
a Gorbea, en el Sur de Chile. La plantación tiene 15 años y ha estado sometida a tres raleos.
}
\examples{
data(eucaplotsampleh2)    
head(eucaplotsampleh2) 
}
\keyword{datasets}
