% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRight.R
\name{extractRight}
\alias{extractRight}
\title{This function extracts the last n-characters of a string.}
\usage{
extractRight(fac, n)
}
\arguments{
\item{fac}{is an object (vector or scalar) of class string or factor}

\item{n}{is the number of characters to be extracted of a the string given in 'fac'.}
}
\value{
This function returns an object having the last-n characters of the object 'fac'
}
\description{
It extracts the last n-characters of a string from the right-hand side.
}
\examples{

last.names.id <- c("Stage-1924","Gregoire-1958","Robinson-1967")
extractRight(last.names.id,5)
years <- extractRight(last.names.id,4)
years
 
}
\author{
Christian Salas-Eljatib.
}
