% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fdamage.r
\docType{data}
\name{fdamage}
\alias{fdamage}
\title{Foliar damage by Ozone}
\format{
Data frame con 52 filas y 2 columnas:
\describe{
\item{damage}{Foliar decoloration, \code{1} with decoloration, \code{0}
without decoloration.}
\item{ozone}{Maximum charge of Ozone concentration.}
}
}
\usage{
fdamage
}
\description{
Foliar damage by Ozone
}
\examples{
data(fdamage)
table(fdamage$damage)
}
\keyword{datasets}
