% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fdamage2.r
\docType{data}
\name{fdamage2}
\alias{fdamage2}
\title{Daño foliar por Ozono}
\format{
Data frame con 52 filas y 2 columnas:
\describe{
\item{damage}{Decoloración foliar, \code{1} con decoloración, \code{0} sin
decoloración.}
\item{ozone}{Máxima carga de concentración de Ozono.}
}
}
\usage{
fdamage2
}
\description{
Daño foliar por Ozono
}
\examples{
data(fdamage2)
table(fdamage2$damage)
}
\keyword{datasets}
