% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{fertiliza2}
\alias{fertiliza2}
\title{Datos a nivel de parcela de un experimento de fertilización}
\format{
Contiene tres columnas como sigue:
\describe{
  \item{tmo}{Tratamiento.Factor medido en diferentes niveles.}
  \item{vol}{Volumen de madera en la parcela experimental, en m\eqn{^{3}}{^3}.}
 }
}
\source{
Datos cedidos por el Prof. Christian Salas.
}
\usage{
data(fertiliza2)
}
\description{
Datos a nivel de parcela de un experimento de fertilización con tratamientos
y replicas.
}
\examples{
data(fertiliza2)    
head(fertiliza2) 
levels(fertiliza$tmo)
media.g <- tapply(fertiliza2$vol,fertiliza2$tmo,mean);media.g
desvst.g <- tapply(fertiliza2$vol,fertiliza2$tmo,sd);desvst.g
n.g <- tapply(fertiliza2$vol,fertiliza2$tmo,length);n.g

}
\references{
not yet
}
\keyword{datasets}
