% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{fertilizaexpe}
\alias{fertilizaexpe}
\title{Fertilization experiment data.}
\format{
Contains three variables, as follows:
\describe{
  \item{treat}{Treatment level.}
  \item{volume}{Plot-level volume, in m\eqn{^{3}}{^3}/plot.}
 }
}
\source{
The data were provided by Dr Christian Salas.
}
\usage{
data(fertilizaexpe)
}
\description{
Data contains volume data at plot-level for a fertilization
experiment.
}
\examples{
data(fertilizaexpe)    
head(fertilizaexpe) 
}
\references{
not yet
}
\keyword{datasets}
