% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{fertilizaexpe2}
\alias{fertilizaexpe2}
\title{Datos a nivel de parcela de un experimento de fertilizacion.}
\format{
Contains three variables, as follows:
\describe{
  \item{treat}{Treatment level.}
  \item{volume}{Plot-level volume, in m3/plot.}
 }
}
\source{
The data were provided by Dr Christian Salas.
}
\usage{
data(fertilizaexpe2)
}
\description{
Data contains volume data at plot-level for a fertilization
experiment.
}
\examples{
data(fertilizaexpe2)    
head(fertilizaexpe2) 
}
\references{
not yet
}
\keyword{datasets}
