% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDataDatana.r
\docType{data}
\name{fishgrowth}
\alias{fishgrowth}
\title{Data on fish growth.}
\format{
Contains 3 variables, as follows:
\describe{
  \item{years}{Year at capture.}
  \item{length}{Length at capture (mm).}
\item{scale}{radius of a key scale (mm).}
 }
}
\source{
The data were obtained from the \eqn{alr4} library of R.
}
\usage{
data(fishgrowth)
}
\description{
Data on samples of small mouth bass collected in West Bearskin Lake, 
Minnesota, in 1991. The
file wblake includes only fish of ages 8 or younger.
}
\examples{
data(fishgrowth)    
head(fishgrowth) 
}
\references{
Weisberg S. 2014. Applied Linear Regression. 4th edition. Hoboken NJ: Wiley
}
\keyword{datasets}
