% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{floraChile}
\alias{floraChile}
\title{Flora of Chile.}
\format{
Contains seven columns, as follows:
\describe{
  \item{family}{.}
  \item{genus}{.}
\item{scientific.name}{.}
\item{author}{.}
\item{origin}{.}
\item{life.form}{.}
\item{lat...}{.}
 }
}
\source{
The data are provided courtesy of Dr Jan Bannister at
 the Instituto Forestal (Chiloe, Chile).
}
\usage{
data(floraChile)
}
\description{
Dataset contains taxonomic level information
segregatted by latitude.
}
\examples{
data(floraChile)    
head(floraChile) 
}
\references{
- Bannister JR, Vidal OJ, Teneb E, Sandoval V. 2012. Latitudinal 
patterns
 and regionalization of plant diversity along a 4270‐km gradient 
 in continental
  Chile. Austral Ecology 37(4):500-509. 
  \doi{10.1111/j.1442-9993.2011.02312.x}
}
\keyword{datasets}
