% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{floraChile2}
\alias{floraChile2}
\title{Flora de Chile.}
\format{
Contains seven columns, as follows:
\describe{
  \item{family}{.}
  \item{genus}{.}
\item{scientific.name}{.}
\item{author}{.}
\item{origin}{.}
\item{life.form}{.}
\item{lat...}{.}
 }
}
\source{
Datos cedidos por el  Dr Jan Bannister del Instituto Forestal (Chiloe, Chile).
}
\usage{
data(floraChile)
}
\description{
Contiene informacion taxonomica segregada por latitude.
}
\examples{
data(floraChile2)    
head(floraChile2) 
}
\references{
- Bannister JR, Vidal OJ, Teneb E, Sandoval V. 2012. Latitudinal patterns
 and regionalization of plant diversity along a 4270‐km gradient in continental
  Chile. Austral Ecology 37(4):500-509.
}
\keyword{datasets}
