% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{football}
\alias{football}
\title{Anaerobic potential of soccer players.}
\format{
The data frame contains 13 variables as follows:
\describe{
  \item{WPM}{Watt Peak Max.}											
\item{WPMk}{Watt Peak Max relative.}
\item{WPm}{Watt Peak media.}
\item{WPmk}{Watt Peak media relative.}
\item{WTT}{Watt Total Work.}
\item{WTTk}{Watt Total Work relative.}
\item{WIF}{Watt Index Fatigue.}
\item{W5}{Watt 5 seg.}
\item{W10}{Watt 10 seg.}
\item{W15}{Watt 15 seg.}
\item{W20}{Watt 20 seg.}
\item{W25}{Watt 25 seg.}
\item{W30}{Watt 30 seg.}
 }
}
\source{
Data were provided by Dr Aquiles Yanez-Silva at Universidad Mayor (Santiago, Chile).
}
\usage{
data(football)
}
\description{
Data about anaerobic variables of football players.
}
\examples{
data(football)    
head(football) 
}
\references{
Not yet.
}
\keyword{datasets}
