% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDataDatana.r
\docType{data}
\name{football2}
\alias{football2}
\title{Potencia anaerobica de jugadores de football.}
\format{
Contiene variables de nivel individual, como se 
describen a continuación::
\describe{
\item{WPM}{}											
\item{WPMk}{}
\item{WPm}{}
\item{WPmk}{}
\item{WTT}{}
\item{WTTk}{}
\item{WIF}{}
\item{W5}{}
\item{W10}{}
\item{W15}{}
\item{W20}{}
\item{W25}{}
\item{W30}{}
 }
}
\source{
Los datos fueron cedidos por el Dr Aquiles Yañez-Silva de 
la Universidad Mayor (Santiago, Chile).
}
\usage{
data(football2)
}
\description{
Datos sobre potencia anaerobica de jugadores de football.
}
\examples{
data(football2)    
head(football2) 
}
\references{
Not yet.
}
\keyword{datasets}
