% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gasoline2.r
\docType{data}
\name{gasoline2}
\alias{gasoline2}
\title{Precios de gasolina y petróleo}
\format{
Data frame que contiene 14 filas y 3 columnas:
\describe{
\item{año}{Año del precio}
\item{gasolina}{Precio de la gasolina para el \code{año} en centavos / galón}
\item{petroleo}{Precio del petróleo para el \code{año} en $ / bbl}
}
}
\source{
McClave, James T. Benson, P.G. 1991. Statistics for Business and
Economics, Fifth Edition. Dellen and Macmillan.
}
\usage{
gasoline2
}
\description{
Precios de gasolina y petróleo
}
\examples{
data(gasoline2)
plot(gasolina~año, data = gasoline2, type = "b",
     ylab = "Precio de la gasolina (centavos/galón)",
     xlab = "Año")
}
\references{
Statistial Abstract of the United States: 1989, pp476, 480.
}
\keyword{datasets}
