% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{hdptorico}
\alias{hdptorico}
\title{Tree height-diameter data of Homalium racemosum in Puerto Rico.}
\format{
Contains three variables, as follows:
\describe{
  \item{tree}{Tree id code.}
\item{dbh}{Diameter at breast height, in cm.}
\item{height}{Total height of the tree, in meters.}
 }
}
\source{
Consider the 40 height-diameter pairs listed in table 1 and 
were typed by Vivian Marschhausen, a student of Prof Christian Salas-Eljatib.
}
\usage{
data(hdptorico)
}
\description{
Height and diameter data for Homalium racemosum trees growing in the Caribbean National Forest on 
the tropical island of Puerto Rico and represent a typical species
 growing in the subtropical wet life zone.
}
\examples{
data(hdptorico)    
head(hdptorico) 
}
\references{
Yuancai L, Parresol B. 2001. Remarks on Height-Diameter Modeling. 
 Res. Note SE-10. U.S. Department of Agriculture, Forest Service,
  Southeastern Forest Experiment Station. 8 p.
}
\keyword{datasets}
