% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{hdptorico2}
\alias{hdptorico2}
\title{Altura y diametro para árboles de Homalium racemosum en Puerto Rico.}
\format{
Contiene tres variables, como sigue:
\describe{
\item{arbol}{Codigo indentificador del árbol.}
\item{dap}{Diametro a la altura del pecho, en cm.}
\item{atot}{Altura total, in meters.}
 }
}
\source{
Los datos provienen del Cuadro No. 1 de Yuancai and Parresol (2001) y 
fueron digitados por Vivian Marschhausen, una estudiante del Prof
 Christian Salas-Eljatib.
}
\usage{
data(hdptorico2)
}
\description{
Altura y diametro medidos en árboles de Homalium racemosum en
 el bosque nacional del caribe en la isla tropical de Puerto Rico y
  representa a una especie tipica de la zona subtropical humeda.
}
\examples{
data(hdptorico2)    
head(hdptorico2) 
}
\references{
Yuancai L, Parresol B. 2001. Remarks on Height-Diameter Modeling. 
 Res. Note SE-10. U.S. Department of Agriculture, Forest Service,
  Southeastern Forest Experiment Station. 8 p.
}
\keyword{datasets}
