% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{hgrowthDfir}
\alias{hgrowthDfir}
\title{Tree height growth of Douglas-fir sample trees in the Northwest of the United States}
\format{
The data frame contains seven variables as follows:
\describe{
  \item{natfor.id}{Code identifier.}
  \item{plot.code}{Plot number identification}
  \item{tree.code}{Tree number identification.}
  \item{dbh}{Diameter at breast height at sampling, in in.}
  \item{htot}{Total height at sa,pling, in ft}
  \item{age}{Age of tree, yr}
  \item{height}{Height at a given age, in ft}
 }
}
\source{
The data were provided by Dr Christian Salas.
}
\usage{
data(hgrowthDfir)
}
\description{
Data contains 148 observations on the height growth of dominant trees of Pseudotsguga mensiezzi in the Northwest of the United States.
}
\examples{
data(hgrowthDfir)    
head(hgrowthDfir)
}
\references{
Monserud RA. Height growth and site index curves for Inland Douglas-
fir based on stem analysis data and forest habitat type. Forest Sci.,
30(4):943-965, 1984.

Salas C, Stage AR, and Robinson AP. Modeling effects of
overstory density and competing vegetation on tree height growth. Forest
Sci., 54(1):107-122, 2008.
}
\keyword{datasets}
