% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{hgrowthDfir2}
\alias{hgrowthDfir2}
\title{Crecimiento en altura de una muestra de arboles en los Estados Unidos}
\format{
La data frame contiene siete variables:
\describe{
  \item{bosque.id}{Codigo identificador del bosque.}
  \item{parcela}{Codigo identificador de la parcela.}
  \item{arbol}{Numero de identificacion arbol.}
  \item{dap}{Diametro a la altura del pecho, en  pulgadas.}
  \item{htot}{Altura total, en pies}
  \item{edad}{Edad, en anhos}
  \item{altura}{Altura para cada edad del arbol, en pies}
 }
}
\source{
La data fue cedida por el Dr Christian Salas-Eljatib.
}
\usage{
data(hgrowthDfir2)
}
\description{
Data contiene 148 obserrvaciones sobre el crecimiento en altura de arboles dominantes de Pseudotsguga mensiezzi en el Nor-Oeste de los Estados Unidos
}
\examples{
data(hgrowthDfir2)    
head(hgrowthDfir2)
}
\references{
Monserud RA. Height growth and site index curves for Inland Douglas-
fir based on stem analysis data and forest habitat type. Forest Sci.,
30(4):943-965, 1984.

Salas C, Stage AR, and Robinson AP. Modeling effects of
overstory density and competing vegetation on tree height growth. Forest
Sci., 54(1):107-122, 2008.
}
\keyword{datasets}
