% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{hgrowthlenga}
\alias{hgrowthlenga}
\title{Height growth of Nothofagus pumilio (lenga) saplings.}
\format{
Data having the following columns:
\describe{
\item{plot}{Sample plot number.}
\item{tree}{Tree code.}
\item{hst}{Stump height, in m.}
\item{year}{Year of establishment.}
\item{age}{Tree age, in yrs.}
\item{height}{Tree height, in m.}
\item{site}{Site code.}
\item{cluster}{Cluster code.}
 }
}
\source{
The data were typed by Valentina Venegas, a student of Prof Christian Salas-Eljatib.
}
\usage{
data(hgrowthlenga)
}
\description{
Tree height growth of lenga saplings in the Ultima Esperanza province,
 on the XII Region (Patagonia) of Chile.
}
\examples{

data(hgrowthlenga)
summary(hgrowthlenga)
}
\references{
Ugalde G. (2006) Crecimiento en altura de renovales de lenga (Nothofagus pumilio (Poepp. et Endl.) Krasser) 
 en Monte Alto(XII Region) en función de la calidad del sitio. Tesis Ingeniero Forestal, Universidad de Chile,
 Santiago, Chile.
}
\keyword{datasets}
