% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{hgrowthlenga2}
\alias{hgrowthlenga2}
\title{Crecimiento en altura de brinzales de Nothofagus pumilio (lenga).}
\format{
Los datos tienen las siguientes columnas:
\describe{
\item{parce}{Numero de la parcela de muestreo.}
\item{arbol}{Codigo del árbol.}
\item{htoc}{Altura de tocon, en m.}
\item{anho}{Año de establecimiento.}
\item{edad}{Edad del árbol, en años.}
\item{altura}{Altura total, en m.}
\item{sitio}{Codigo del sitio.}
\item{conglo}{Codigo del conglomerado.}
 }
}
\source{
Los datos fueron digitados por Valentina Venegas, una estudiante del Prof Christian Salas-Eljatib.
}
\usage{
data(hgrowthlenga2)
}
\description{
Datos de crecimiento en brinzales de lenga en la provincia de Ultima Esperanza province,
 en la XII Region (Patagonia) de Chile.
}
\examples{

data(hgrowthlenga2)
summary(hgrowthlenga2)
}
\references{
Ugalde G. (2006) Crecimiento en altura de renovales de lenga (Nothofagus pumilio (Poepp. et Endl.) Krasser) 
 en Monte Alto(XII Region) en función de la calidad del sitio. Tesis Ingeniero Forestal, Universidad de Chile,
 Santiago, Chile.
}
\keyword{datasets}
