% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-leafw2.r
\docType{data}
\name{leafw2}
\alias{leafw2}
\title{Area y peso para hojas de árboles.}
\format{
Contiene variables a nivel de hoja, como sigue:
\describe{
\item{peso}{Peso de la hoja, en gramos.}
\item{area}{Área foliar, en cm\eqn{^{2}}{^2}.}
}
}
\source{
El archivo de datos fue cortesía del Prof. Timothy
Gregoire de Yale University (New Haven, CT, USA).
}
\usage{
data(leafw2)
}
\description{
Mediciones de área y peso de hojas.
}
\examples{
library(datana)
data(leafw2)    
head(leafw2)
plot(peso~area, data=leafw2) 
}
\references{
\itemize{
\item Gove JH, Barrett JP, and Gregoire TG. 1982. When is n sufficiently large
for regression estimation? Journal of Environmental Management 15:229-237.
}
}
\keyword{datasets}
