% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{lleuque}
\alias{lleuque}
\title{Contains species composition data of Prumnopitys andina (Lleuque) forests}
\format{
The dataframe has the following columns
\describe{
  \item{stand}{Stand number}
  \item{plot.num}{Plot number}
  \item{Aus.chi}{Tree density/ha of Austrocedrus chilensis}
  \item{May.dis}{Tree density/ha of Maytenus disticha}
  \item{Not.obl}{Tree density/ha of Nothofagus obliqua}
  \item{Pru.and}{Tree density/ha of Prumnopitys andina}
 }
}
\source{
The data are provided courtesy of Prof. Rodrigo Vargas-Gaete at Universidad
 de La Frontera (Temuco, Chile).
}
\usage{
lleuque
}
\description{
Contains species composition data for forests with presence of Lleuque (Prumnopitys andina)
}
\references{
Vargas-Gaete R, Salas-Eljatib C, Penneckamp D, Neira Z, Diez MC, Vargas-Picón, R. 2020. Estructura 
y regeneración de bosques de Prumnopitys andina en los Andes del sur de Chile. Gayana Botánica, 77(1), 48-58.
}
\keyword{datasets}
