% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDataDatana.r
\docType{data}
\name{lleuque2}
\alias{lleuque2}
\title{Densidad por especie de árboles por parcela en bosques de 
*Prumnopitys andina* (Lleuque)}
\format{
Los datos contienen las siguientes columnas:
\describe{
  \item{rodal}{Rodal, en código.}
  \item{parce}{Parcela de muestreo, en código.}
  \item{cipres}{Densidad de Ciprés de la Cordillera
   (Austrocedrus chilensis), en arb/ha.}
  \item{ldura}{Densidad de leña dura (Maytenus disticha), en arb/ha.}
  \item{roble}{Densidad de roble (Nothofagus obliqua), en arb/ha.}
  \item{lleuque}{Densidad de lleuque (Prumnopitys andina), en arb/ha.}
 }
}
\source{
Los datos fueron cedidos por el Prof. Rodrigo Vargas-Gaete de la Universidad
 de La Frontera (Temuco, Chile).
}
\usage{
lleuque2
}
\description{
Para cada unidad de muestreo (o parcela) hay valores de densidad 
por especie arbórea. Las parcelas
fueron establecidas en bosques de *Prumnopitys andina* (Lleuque) forests
}
\references{
Vargas-Gaete R, Salas-Eljatib C, Penneckamp D, Neira Z, Diez MC,
 Vargas-Picón, R. 2020. Estructura 
y regeneración de bosques de Prumnopitys andina en los Andes del 
sur de Chile. Gayana Botánica
 77(1):48-58. \doi{10.4067/S0717-66432020000100048}
}
\keyword{datasets}
