% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.r
\name{moda}
\alias{moda}
\title{Computes the mode}
\usage{
moda(y = y)
}
\arguments{
\item{y}{is a numeric vector.}
}
\value{
The function returns the mode, a numeric scalar.
}
\description{
Computes the mode of a random variable.
}
\details{
The mode is an statistics representing the most "used"
value of a random variable as a measurement of central position.
We use the Spanish name of mode, i.e., "moda", to avoid any
confution with the \code{mode} function of R, wich was programmed
for a different use.
}
\examples{

library(datana)
data(casen)
head(casen)
df<-casen
#Compare 
mean(df$edad)
median(df$edad)
# Using the function
moda(df$edad)
}
\author{
Christian Salas-Eljatib.
}
