% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDataDatana.r
\docType{data}
\name{papersdocstu}
\alias{papersdocstu}
\title{Productividad científica de estudiantes de postgrado}
\format{
Este set de datos contiene las siguientes columnas:
\describe{
  \item{papers}{Es el número de artículos cientificos publicados
  luego de 3 años de egresado.}
  \item{genero}{Hombre/mujer.}
  \item{est.civil}{Estado civil del egresado.}
  \item{nin.men5}{Número de hijos menores a 5 años que dependen del
  egresado.}
  \item{prog.prest}{Puntaje asignado al prestigio del programa de 
  postgrado.}
  \item{papers.guia}{Número de papers publicados por el profesor(a)
  guía del egresado, en el mismo periodo de tiempo.}
 }
}
\source{
Los datos fueron obtenidos desde el paquete 'pslp' de R
}
\usage{
data(papersdocstu)
}
\description{
Corresponde a un estudio realizado en xxxx, sobre el numero
de papers publicados por estudiantes egresados de programas
de doctorado luego de 3 años.
}
\examples{
data(papersdocstu)
df<-papersdocstu    
head(df) 
table(df$genero)
table(df$est.civil,df$genero)
tapply(df$papers,df$est.civil,summary)
}
\keyword{datasets}
