% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pesohojas.r
\docType{data}
\name{pesohojas}
\alias{pesohojas}
\title{Peso de hojas}
\format{
Data frame con 64 filas y 2 columnas:
\describe{
\item{peso}{peso foliar en gramos (g)}
\item{area}{área foliar en centímetros cuadrados (cm²)}
}
}
\usage{
pesohojas
}
\description{
Peso de hojas
}
\examples{
data(pesohojas)
plot(peso~area, data = pesohojas)
}
\keyword{datasets}
