% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{populavolplots2}
\alias{populavolplots2}
\title{Una población de parcelas de muestreo con datos de volumen.}
\format{
Los datos contienen una sola columna:
\describe{
\item{y}{Volumen agregado de la parcela, en m\eqn{^{3}}{^3}. En este caso
 el volumen en m\eqn{^{3}}{^3}/ha se obtendria al multiplicarlo por el factor
 de expansion (100).}
 }
}
\source{
Los datos fueron digitados por Natalia Pizarro, Javiera Zuñiga y Jose Osses, todos ellos estudiantes del Profesor Christian Salas-Eljatib.
}
\usage{
data(populavolplots2)
}
\description{
Volumen agregado de árboles en 576 parcelas contiguas de 10m*10m. Por lo tanto, la poblacion
 cubre 5.76 ha. Los datos provienen de una plantación de Pinus radiata de 22 años
 ubicada a 15 Km al NorOeste de Valdivia, Chile.
}
\examples{

data(populavolplots2)
summary(populavolplots2$volumen)
hist(populavolplots2$volumen)
}
\references{
Wrann JH. 1970. Efectividad de los distintos sistemas de muestreo y metodos de
 calculo en el inventario de un rodal de pino insigne. Tesis Ingeniero Forestal, Universidad Austral de Chile,
 Valdivia, Chile. 51p.
}
\keyword{datasets}
