% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{presidentChile}
\alias{presidentChile}
\title{2021 presidential election in Chile.}
\format{
The dataframe contains the following columns:
\describe{
  \item{region.no}{Administrative region number of Chile.}
  \item{region}{Administrative region name.}
  \item{provincia}{Province.}
  \item{circu.senatorial}{Senatorial constituency.}
  \item{distrito}{Distrit.}
  \item{comuna}{County.}
  \item{region}{Senatorial constituency.}
  \item{provincia}{Province.}
  \item{circu.elec}{Electoral constituency.}
  \item{local}{Place.}
  \item{no.mesa}{Voting table.}
  \item{tipo.mesa}{Voting table type.}
  \item{mesas.fusionadas}{Merged voting tables.}
  \item{electores}{Electors.}
  \item{nro.en.voto}{.}
  \item{candidato}{Candidate. Gabriel Boric and Jose A. Kast}
  \item{votos.tricel}{Total number of votes.}       
 }
}
\source{
The data were obtained from the electoral service of the Chilean Government (SERVEL) at
 \url{https://www.servel.cl/resultados-definitivos-elecciones-segunda-votacion-presidencia-2021/}.
  The datafile name was "Resultados_mesa_presidencial_TRICEL_2v_2021-1.xlsx", and was downloaded
  on October 24, 2022.
}
\usage{
data(presidentChile)
}
\description{
Voting table-level data of the 2021 presidential election in Chile. The election was held on December 19, 2021.
}
\examples{
data(presidentChile)    
head(presidentChile) 
}
\keyword{datasets}
