% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{presidentChile2}
\alias{presidentChile2}
\title{2021 presidential election in Chile.}
\format{
The data frame contains the following columns:
\describe{
  \item{region.no}{Administrative region number of Chile.}
  \item{region}{Administrative region name.}
  \item{provincia}{Province.}
  \item{circu.senatorial}{Circunscripción senatorial.}
  \item{distrito}{Distrito.}
  \item{comuna}{Comuna.}
  \item{circu.elec}{Circunscripción electoral.}
  \item{local}{Lugar de votacion.}
  \item{no.mesa}{Numero de mesa de votacion.}
  \item{tipo.mesa}{Tipo de mesa: M (Mujeres), V (Varones).}
  \item{mesas.fusionadas}{Mesas funcionadas.}
  \item{electores}{Total de electores de la mesa.}
  \item{nro.en.voto}{Numero del candidato en el voto.}
  \item{candidato}{Candidato: Gabriel Boric, Jose A. Kast, Votos nulos, Votos en blanco.}
  \item{votos.tricel}{Numero total de votos de la mesa}       
 }
}
\source{
Los datos fueron obtenidos desde el servicio electoral del Gobierno de Chile (SERVEL) en
 \url{https://www.servel.cl/centro-de-datos/procesos-electorales-detalle/?id=27}.
  El archivo de datos tenia el nombre "Resultados_mesa_presidencial_TRICEL_2v_2021-1.xlsx", y fue 
  descargado el 24 de Octubre del 2022.
}
\usage{
data(presidentChile)
}
\description{
Voting table-level data of the 2021 presidential election in Chile. The election was held on December 19, 2021.
}
\examples{
data(presidentChile2)    
head(presidentChile2) 
}
\keyword{datasets}
