% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{primary}
\alias{primary}
\title{2021 primary election for the president of Chile.}
\format{
The dataframe contains the following columns:
\describe{
  \item{region.no}{Administrative region number of Chile.}
  \item{region}{Administrative region name.}   
  \item{province}{Province.}
  \item{distrit}{Distrit.}
  \item{county}{County.}
  \item{elect.const}{Electoral constituency.}
  \item{location}{Location.}
  \item{table.type}{Voter table type.}
  \item{table}{Voter table.}
  \item{merged.tab}{Merged voter tables.}
  \item{num.in.vote}{.}
  \item{list}{Specific political list.}
  \item{pact}{Specific political pact.}
  \item{party}{Political party.}
  \item{candidate}{Candidate name}
  \item{total.votes}{Total number of votes.}       
 }
}
\source{
The data were obtained from the electoral service of the Chilean Government (SERVEL) at
 \url{https://www.servel.cl}.
  The datafile name was "Resultados_Primarias_Presidenciales_2021_CHILE.xlsx", and was downloaded
  on October 24, 2022.
}
\usage{
data(primary)
}
\description{
Voter table-level data of the 2021 primary election for the president of Chile.
}
\examples{
data(primary)    
head(primary) 
}
\keyword{datasets}
