% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{primary2}
\alias{primary2}
\title{Elección primaria para la presidencia de Chile}
\format{
The dataframe contains the following columns:
\describe{
  \item{region.no}{Región administrativa de Chile.}
  \item{region}{Administrative region name.}   
  \item{provincia}{Provincia.}
  \item{distrito}{Distrito.}
  \item{comuna}{Comuna.}
  \item{circu.elec}{Circunscripcion electoral.}
  \item{local}{Local de votacion.}
  \item{tipo.mesa}{Voter table type.}
  \item{mesa}{Voter table.}
  \item{mesas.fusionadas}{Merged voter tables.}
  \item{nro.voto}{.}
  \item{lista}{Lista politica del candidato.}
  \item{pacto}{Pacto politico del candidato.}
  \item{partido}{Partido politica del candidato.}
  \item{candidato}{Nombre del candidato.}
  \item{votos}{Número total de votos.}       
 }
}
\source{
The data were obtained from the electoral service of the Chilean Government (SERVEL) at
 \url{https://www.servel.cl}.
  The datafile name was "Resultados_Primarias_Presidenciales_2021_CHILE.xlsx", and was downloaded
  on October 24, 2022.
}
\usage{
data(primary2)
}
\description{
Datos a nivel de mesa para la votacion primaria de presidente de Chile.
}
\examples{
data(primary2)    
head(primary2) 
}
\keyword{datasets}
