% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pspruca2.r
\docType{data}
\name{pspruca2}
\alias{pspruca2}
\title{Ubicación espacial de árboles en el bosque de Rucamanque}
\format{
Las columnas describen características de los árboles vivos en
pie, como sigue:
\describe{
\item{arbol}{Número del árbol}
\item{spp}{Codificación de la especie como sigue:
"A. punctatum" es \emph{Aextoxicon punctatum},
"E. cordifolia" es \emph{Eucryphia cordifolia},
"G. avellana" es \emph{Gevuina avellana},
"L. dentata" es \emph{Lomatia dentata},
"L. philippiana" es \emph{Laureliopsis philippiana},
"L. sempervirens" es \emph{Laurelia sempervirens},
"N. obliqua" es \emph{Nothofagus obliqua}, "P. lingue" es \emph{Persea lingue},
y "Other" representa a cualquier especie distinta a cualquiera
de las ya definidas.}
\item{clase.copa}{Clase de copa (1: superior, 2: intermedio, 3; inferior)}
\item{dap}{Diámetro a la altura del pecho, en cm}
\item{coord.x}{Posicion cartesiana en el eje X, en m}
\item{coord.y}{Posicion cartesiana en el eje Y, en m}
}
}
\source{
Los datos fueron cedidos por el Dr Christian Salas-Eljatib (Santiago, Chile).
}
\usage{
data(pspruca2)
}
\description{
Medidas a nivel de árbol y coordenadas espaciales en un parcela de
muestreo permanente de 1 ha (100 x 100m) en el
bosque de Rucamanque, cerca de Temuco, Chile. Mayores antecedentes en
las referencias.
}
\examples{
data(pspruca2)
table(pspruca2$spp)
}
\references{
Salas C, LeMay V, Nunez P, Pacheco P, and Espinosa A. 2006.
Spatial patterns in an old-growth \emph{Nothofagus obliqua} forest in
south-central Chile. Forest Ecology and Management 231(1-3): 38-46.
\doi{10.1016/j.foreco.2006.04.037}
}
\keyword{datasets}
