% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDataDatana.r
\docType{data}
\name{ptaeda2}
\alias{ptaeda2}
\title{Crecimiento en altura de Pinus taeda}
\format{
Los datos contienen las siguientes columnas:
\describe{
\item{semilla.id}{Un factor indicando el origen de la semilla del árbol.}
\item{edad}{Edad del árbol, en años.}
\item{atot}{Altura total, en m.}
 }
}
\source{
Pinheiro, J. C. and Bates, D. M. (2000) Mixed-effects Models in S and 
S-PLUS. Springer.
}
\usage{
data(ptaeda2)
}
\description{
Esta dataframe contiene 84 folas y tres columnas de crecimiento en altura de 
árboles de Pinus taeda (Loblolly pine). Es una modificación de la dataframe
 "Loblolly" del paquete 'datasets' de R.
}
\examples{

data(ptaeda2, package="datana")
head(ptaeda2)
plot(atot ~ edad, data = subset(ptaeda2, semilla.id == 329),
     xlab = "Edad (años)", las = 1,
     ylab = "Altura (m)")
}
\keyword{datasets}
