% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{radiatapl}
\alias{radiatapl}
\title{Sampling plots data from a Pinus radiata plantation near Capitan Pastene, 
Region de La Araucania, Chile.}
\format{
The data frame contains four variables as follows:
\describe{
  \item{plot}{Plot number identification.}
  \item{tree}{Tree number identification.}
  \item{dbh}{Diameter at breast height, in cm.}
  \item{heigth}{Total height, in m.}
 }
}
\source{
The data are provided courtesy of Mr. Mauricio Lobos-Beneventi 
(Temuco, Chile).
}
\usage{
data(radiatapl)
}
\description{
Tree-level information collected within sample plots in a forestry 
plantation of Pinus radiata near Capitan Pastene, in southern Chile. 
Sample plots size is 150 m\eqn{^{2}}{^2}.
}
\examples{
data(radiatapl)    
head(radiatapl) 
}
\keyword{datasets}
