% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{radiatapl2}
\alias{radiatapl2}
\title{Datos a nivel de árbol de parcelas de muestreo en plantaciones de Pinus radiata}
\format{
Los datos contienen las siguientes columnas
\describe{
  \item{parce}{Número de identificación de la parcela de muestreo.}
  \item{arbol}{Número de identificación del árbol dentro de la parcela.}
  \item{dap}{Diámetro a los 1.3 m en el fuste, en cm.}
  \item{atot}{Altura total, en m. Solo registrada para algunos árboles muestra.}
 }
}
\source{
Los datos son cortesia del Ing. Forestal Mauricio Lobos-Beneventi (Temuco, Chile).
}
\usage{
data(radiatapl2)
}
\description{
Es un listado de árboles con características medidas dentro de unidades de muestreo en una plantacion de Pinus radiata cercana a Capitan Pastene, 
Region de la Araucania, Chile. Las parcelas de muestreo tienen 150 m\eqn{^{2}}{^2}.
}
\examples{
data(radiatapl2)    
head(radiatapl2) 
}
\keyword{datasets}
