% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rainfallCA.r
\docType{data}
\name{rainfallCA}
\alias{rainfallCA}
\title{Datos de precipitaci贸n en Californa}
\format{
Este set de datos contiene las siguientes columnas:
\describe{
\item{saple.id}{Identificador del punto de muestreo.}
\item{easting}{Coordenada este del punto.}
\item{northing}{Coordenada norte del punto.}
\item{pp}{Precipitaci贸n, en pulgadas.}
\item{ele}{Elevaci贸n, en pies.}
\item{lat}{Latitud del punto.}
\item{d.coast}{Distancia a la costa, en millas.}
}
}
\source{
Los datos provienen de mediciones hechas en California
}
\usage{
data(rainfallCA)
}
\description{
Datos de precipitaci贸n medidos en distintos lugares de california,
con las coordenadas de los puntos y su distancia a la costa.
}
\examples{
data(rainfallCA)    
head(rainfallCA) 
plot(pp~ele, data=rainfallCA)
hist(rainfallCA$pp)
}
\keyword{datasets}
