% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{raulihg}
\alias{raulihg}
\title{Height growth of Nothofagus alpina trees in Chile.}
\format{
The data frame contains four variables as follows:
\describe{
\item{tree.code}{Tree id code}											
\item{spp}{Species common name}
\item{bha.t}{Breast-height age, in yrs.}
\item{h.t}{Total height, in m.}
 }
}
\source{
Data were provided by Dr Christian Salas-Eljatib.
}
\usage{
data(raulihg)
}
\description{
Time series data of height for Nothofagus alpina (rauli) trees in south-central Chile. These sampled trees are part of the ones
 used in the cited article (reference section).
}
\examples{
data(raulihg)    
head(raulihg) 
}
\references{
- Salas-Eljatib C. 2021. An approach to quantify climate-productivity relationships: an
 example from a widespread Nothofagus forest. Ecological Applications 31(4): e02285.

- Salas-Eljatib, C. 2021. Time series height-data for Nothofagus alpina trees. \doi{10.6084/m9.figshare.13521602.v5}
}
\keyword{datasets}
