% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{redoakvol}
\alias{redoakvol}
\title{Contains tree-level merchantable volume data for red oak (Quercus rubra).}
\format{
\describe{
  \item{tree}{Tree number.}
  \item{dbh}{Diameter at breast height, in cm.}
  \item{merc.h}{Merchantable height to a 4.5 inch dob, in m.}
  \item{merc.vol}{Merchantable peeled volume to a  minimum 4.5-inch outside bark diameter, in m\eqn{^{3}}{^3}.}
 }
}
\source{
The data are provided courtesy of Prof. Timothy Gregoire at
the School of Forestry and Environmental Studies at Yale University (New Haven, CT, USA).
}
\usage{
redoakvol
}
\description{
Merchantable volume of 236 sample trees of Red oak from Pennsylvania (USA).
}
\examples{
head(redoakvol) 
}
\references{
- Beers TW, Gingrich SF. 1958. Construction of Cubic-foot Volume Tables for Red Oak in
 Pennsylvania. Journal of Forestry 56(3):210-214.

- Moser JW, Beers TW. 1969. Parameter Estimation in Nonlinear Volume
 Equations. Journal of Forestry 67(12):878-879.
}
\keyword{datasets}
