% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{redoakvol2}
\alias{redoakvol2}
\title{Volumen comercial de 치rboles de red oak (Quercus rubra).}
\format{
\describe{
  \item{arbol}{Numero del 치rbol.}
  \item{dap}{Diametro a la altura del pecho, en cm.}
  \item{hiu}{Altura comercial donde el fuste tiene 4.5 pulgadas con corteza, en m.}
  \item{vol.iu}{Volumen comercial a un indice de utilizaci칩n de 4.5 pulgadas de diametro con corteza, in m\eqn{^{3}}{^3} sin corteza.}
 }
}
\source{
The data are provided courtesy of Prof. Timothy Gregoire at
the School of Forestry and Environmental Studies at Yale University (New Haven, CT, USA).
}
\usage{
redoakvol2
}
\description{
Volumen comercial de 236 치rboles muestra de Quercus rubra (red oak) en el estado de Pennsylvania (USA).
}
\examples{
head(redoakvol2) 
}
\references{
- Beers TW, Gingrich SF. 1958. Construction of Cubic-foot Volume Tables for Red Oak in
 Pennsylvania. Journal of Forestry 56(3):210-214.

- Moser JW, Beers TW. 1969. Parameter Estimation in Nonlinear Volume
 Equations. Journal of Forestry 67(12):878-879.
}
\keyword{datasets}
