% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{simula}
\alias{simula}
\title{Simulated yield of forestry plantations of exotic species in Chile.}
\format{
Contains stand-level variables, as follows:
\describe{
  \item{species}{"P. radiata" is Pinus radiata, "E. globulus" 
  is Eucalyptus globulus, and "E. nitens" is
  Eucalyptus nitens.}
\item{age}{plantation age, in years.}
  \item{tph }{Tree density, in trees/ha.}
 \item{gha }{Basal area, in m\eqn{^{2}}{^2}/ha.}
\item{toph}{Dominant height, in m.}
 \item{qmd}{quadratic mean diameter, in cm.}
\item{totvol}{gross stand volume, in m\eqn{^{3}}{^3}/ha} 
\item{viu.10}{stand volume below an utilizacion index of 10 cm,
 in m\eqn{^{3}}{^3}/ha.} 
\item{viu.15}{stand volume below an utilizacion index of 15 cm, 
in m\eqn{^{3}}{^3}/ha.} 
\item{viu.20}{stand volume below an utilizacion index of 20 cm, 
in m\eqn{^{3}}{^3}/ha.} 
\item{viu.25}{stand volume below an utilizacion index of 25 cm, 
in m\eqn{^{3}}{^3}/ha.} 
 }
}
\source{
The data were obtained as outputs for plantations without
  management in Chile. The academic
 version of the simulator was used. You can visit mnssimulacion.cl
}
\usage{
data(simula)
}
\description{
The yield tables of simulated plantations of Pinus radiata,
 Eucalyptus globulus, and Eucalyptus nitens
 are obtained from the Radiata simulator and EucaSim simulator built
  in Chile. Several stand-level variables 
 are part of the output.
}
\examples{
data(simula)    
}
\keyword{datasets}
