% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleupp.r
\name{singleupp}
\alias{singleupp}
\title{Convert}
\usage{
singleupp(fac)
}
\arguments{
\item{fac}{is an object of class string or factor}
}
\value{
This function returns an object having the resulting
string.
}
\description{
Function to get the first letter only as upper case of a string.
}
\details{
It is specially set to arrange an data vector having
alphanumeric format.
}
\examples{
singleupp("PETER")
cities <- c("bogota","lima","new york","santiago","madrid")
singleupp(cities)
}
\references{
Salas-Eljatib, C. 2021. Análisis de datos con el programa
estadístico R: una introducción
aplicada. Ediciones Universidad Mayor, Santiago, Chile. 170 p.
\url{https://eljatib.com/rlibro}
}
\author{
Christian Salas-Eljatib
}
