% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{slashpine}
\alias{slashpine}
\title{Biomass dataset}
\format{
The data frame contains nine variables as follows:
\describe{
\item{tree_id}{tree code}											
\item{dbh}{diameter}
\item{h}{heigth}
\item{lcl}{live crown lenght}
\item{age}{age tree}
\item{wood}{wood biomass}
\item{bark}{bark biomass}
\item{crown}{crown biomass}
\item{tree}{tree biomass}
 }
}
\source{
Data were provided by Dr Christian Salas-Eljatib (Universidad Mayor, Santiago, Chile).
}
\usage{
data(slashpine)
}
\description{
Dataset that contains nine pairs of columns with information about biomass of 40 samples.
}
\examples{
data(slashpine)    
head(slashpine) 
}
\references{
Parresol BR. 2001. Additivity of nonlinear biomass equations. Canadian Journal of For Research, 31:865-878.
}
\keyword{datasets}
