% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{slashpine2}
\alias{slashpine2}
\title{Biomasa}
\format{
Variables se describen a continuacion::
\describe{
\item{tree_id}{Identifcador árbol}											
\item{dbh}{diametro}
\item{h}{altura total}
\item{lcl}{largo de copa}
\item{age}{edad árbol}
\item{wood}{biomasa madera}
\item{bark}{biomasa corteza}
\item{crown}{biomasa copa}
\item{tree}{biomasa árbol}
 }
}
\source{
Datos fueron contribuidos por el Dr Christian Salas-Eljatib (Universidad Mayor, Santiago, Chile).
}
\usage{
data(slashpine2)
}
\description{
Dataset que contiene 9 pares de columnas con información sobre la biomasa de 40 árboles.
}
\examples{
data(slashpine2)    
head(slashpine2) 
}
\references{
Parresol BR. 2001. Additivity of nonlinear biomass equations. Canadian Journal of For Research, 31:865-878.
}
\keyword{datasets}
