% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{sludge}
\alias{sludge}
\title{Sludge data are at different cities, with a value of concentration zinc.}
\format{
Contains four variables, as follows:
\describe{
  \item{city}{Name of city.}
  \item{rate}{Concentration rate of sludge.}
\item{zinc}{Value of concentration ( in ppm).}
\item{trt.comb}{Combination between city and rate factors.}
 }
}
\source{
The data were provided from.
}
\usage{
data(sludge)
}
\description{
Dataset contains 36 observations
}
\examples{
data(sludge)    
head(sludge) 
}
\references{
not yet
}
\keyword{datasets}
