% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{snaspeChile2}
\alias{snaspeChile2}
\title{Sistema nacional de areas protegidas del estado (SNASPE) de Chile}
\format{
Contiene las siguientes variables para cada unidad del SNASPE:
\describe{
\item{uni.id}{número indentificador de la unidad.}
\item{unidad}{Nombre de la unidad.}
\item{categoria}{Categoría de la unidad. It can be either a National Park, a National Reserve or a Natural Monument.}
\item{comuna}{Name of the commune (the smallest Chilean territorial division) where the unit is located.}
\item{province}{Province where the comunne is located (one territorial division level above the commune).}
 \item{region}{Region where the province is located (one territorial division level above the province and the biggest Chilean territorial division).}
\item{perim.km}{Perimeter, in km.}
\item{area.ha}{Area, in hectares.}
\item{area.m2}{Area, in m\eqn{^{2}}{^2}.}
 }
}
\source{
Estos datos fueron obtenidos desde 
\url{https://ide.minagri.gob.cl/geoweb/2019/11/21/medio-ambiente/}
}
\usage{
data(snaspeChile2)
}
\description{
Contiene variables general de las unidades del sistema de areas protegidas por el estado de 
 Chile (SNASPE).
}
\examples{
data(snaspeChile2)    
head(snaspeChile2) 
}
\references{
EL SNASPE esta bajo la administracion de la Corporacion Nacional Forestal (CONAF) de Chile.
Mayor informacion y documentacion se puede encontrar en
\url{https://www.conaf.cl}
}
\keyword{datasets}
