% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{soiltreat2}
\alias{soiltreat2}
\title{Tratamiento de suelo en el crecimiento de plantulas de  2-años de edad.}
\format{
Contiene las siguientes columnas:
\describe{
  \item{bloque}{Codigo del bloque correspondiente del experimento.} 
  \item{tmo}{Tratamiento.}
  \item{alt.ini}{Altura inicial de la parcela, en m.}
  \item{alt.inc}{Incremento en altura de los ultimos 5-años de la parcela, en m.}
 }
}
\source{
Los datos vienen de la pagina 71 de Freese (1967), y fueron digitados por Nayeli
 Ramirez,  una estudiante del Prof Christian Salas-Eljatib.
}
\usage{
data(soiltreat)
}
\description{
Los tratamientos fueron asignados aleatoriamente a tres parcelas dentro de cada uno de los 11 bloques.
 Cada parcela estaba constituida de 50 plantulas. El incremento promedio en altura de los ultimos 5 años
  fue el criterio para evaluar los tratamientos. La altura inicial y el incremento son las variables continuas
  de los datos.
}
\examples{
data(soiltreat2)    
tapply(soiltreat2$alt.inc,soiltreat2$tmo,summary)
}
\references{
Freese F. 1967. Elementary statistical methods for foresters. Agriculture Handbook 317,
U.S. Department of Agriculture, Forest Service, Washington DC, USA. 87p.
}
\keyword{datasets}
