% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDataDatana.r
\docType{data}
\name{sppTraits}
\alias{sppTraits}
\title{Contains information of functional traits of species.}
\format{
Contains 17 variables, as follows:
\describe{
	\item{sp}{Abbreviated name of specie.}
  	\item{sp.name}{Name of specie.}
	\item{family}{Family of specie.}
	\item{genus}{Genus of specie.}
	\item{phyl}{Type of phylogeny.}
	\item{l.hab}{Type of leaf habit.}
	\item{leaf}{Type of leaf.}
	\item{lt}{.}
	\item{lma}{Leaf mass area.}
	\item{amass}{Photosynthetic capacity per unit leaf mass.}
	\item{n.mass}{Leaf N content per unit mass.}
	\item{pmass}{Leaf P content per unit mass.}
	\item{l.lifespan}{Leaf life span.}
	\item{l.length}{Leaf length.}
	\item{sem}{Seed mass.}
	\item{wd}{Wood density.}
	\item{max.h}{Maximum height.}
  }
}
\source{
The data were provided from DRYAD repository
}
\usage{
data(sppTraits)
}
\description{
Dataset contains 48 observations about about functional trait values 
for each of the 48 study species, 
including 23 evergreen and 25 deciduous.
}
\examples{
data(sppTraits)    
head(sppTraits) 
}
\references{
- Ameztegui A, Paquette A, Shipley B, Heym M, Messier C, Gravel D. 2016. 
 Shade tolerance and the functional trait: demography relationship 
 in temperate and boreal forests. Functional Ecology 
 31: 821-830.
}
\keyword{datasets}
