% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{standLleuque}
\alias{standLleuque}
\title{Plot-level data with variables from Andean Prumnopitys forests}
\format{
The data frame contains seven variables as follows:
\describe{
\item{rodal}{number of stand}											
\item{plot.id}{code of plot}
\item{nha}{Density of plot}
\item{gha}{Basal area of plot}
\item{qmd}{Quadratic mean diameter of plot}
\item{toph}{Dominant height of plot}
\item{structure}{Forest structure level: open, secondary adult, pure}
 }
}
\source{
The data are provided courtesy of Prof. Rodrigo Vargas-Gaete at Universidad
 de La Frontera (Temuco, Chile).
}
\usage{
data(standLleuque)
}
\description{
Data on density, basal area, mean square diameter and other 
variables of 24 plots for Lleuque is provided.
}
\examples{
data(standLleuque)    
head(standLleuque) 
}
\references{
Vargas-Gaete R, Salas-Eljatib C, Penneckamp D, Neira Z, Diez MC, 
Vargas-Picón, R. 2020. Estructura 
y regeneración de bosques de Prumnopitys andina en los Andes del sur
 de Chile. Gayana Botánica
 77(1):48-58. \doi{10.4067/S0717-66432020000100048}
}
\keyword{datasets}
