% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{standLleuque2}
\alias{standLleuque2}
\title{Datos con variables a nivel de parcela de bosques de Prumnopitys andina}
\format{
Variables se describen a continuacion::
\describe{
\item{rodal}{Numero de rodal}											
\item{plot.id}{Codigo de parcela}
\item{nha}{Densidad de parcela}
\item{gha}{Area basal de parcela}
\item{qmd}{Diametro medio cuadratico}
\item{toph}{Altura dominante}
\item{estructura}{Estructura del bosque. Abiero, secundario adulto o puro}
 }
}
\source{
The data are provided courtesy of Prof. Rodrigo Vargas-Gaete at Universidad
 de La Frontera (Temuco, Chile).
}
\usage{
data(standLleuque2)
}
\description{
Se proporciona informacion de densidad, area basal, diametro medio cuadratico  y otras variables de 24 parcelas para Lleuque.
}
\examples{
data(standLleuque2)    
head(standLleuque2) 
}
\references{
Vargas-Gaete R, Salas-Eljatib C, Penneckamp D, Neira Z, Diez MC, Vargas-Picón, R. 2020. Estructura 
y regeneración de bosques de Prumnopitys andina en los Andes del sur de Chile. Gayana Botánica, 77(1), 48-58.
}
\keyword{datasets}
