% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{traits}
\alias{traits}
\title{Functional traits of vegetative species in Chile.}
\format{
\describe{
  \item{esp}{species codified name}
  \item{shadeTolerance}{indicates the species tolerance to shape. There are three main classes: shade-tolerant, shade-midtolerant and shade-intolerant}
  \item{spp.ci.name}{Scientific name.}
  \item{spp.ci.abb.}{.}
  \item{wd}{wood density in kg per cubic meters.}
 }
}
\source{
Some of the information on shade tolerance can be found in Soto et al 2010.
}
\usage{
data(traits)
}
\description{
Functional traits of vegetative species in Chile. Includes column with codified name (esp)
}
\references{
- Soto DP, Salas C, Donoso PJ, Uteau D. 2010. Heterogeneidad estructural y espacial de un bosque mixto dominado por
Nothofagus dombeyi despues de un disturbio parcial. Revista Chilena de Historia Natural 83(3): 335-347.
}
\keyword{datasets}
