% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{traits2}
\alias{traits2}
\title{Rasgos funcionales para algunas especies vegetales de Chile.}
\format{
\describe{
  \item{especie}{Codigo alfanumerico para especie.}
  \item{tolerancia.sombra}{Tolerancia a la sombra de la especie.}
  \item{nombre.cient}{Nombre cientifico.}
  \item{nom.cient.abre}{Nombre cientifico abreviado.}
  \item{den.madera}{Densidad de la madera en kg/m\eqn{^{3}}{^3}.}
 }
}
\source{
Parte de la informacion sobre tolerancia a la sombra se encuentra en 
Soto et al 2010
}
\usage{
data(traits2)
}
\description{
Rasgos funcionales para algunas especies vegetales de Chile.
}
\references{
- Soto DP, Salas C, Donoso PJ, Uteau D. 2010. Heterogeneidad estructural y
 espacial de un bosque mixto dominado por
Nothofagus dombeyi despues de un disturbio parcial. Revista Chilena de 
Historia Natural 83(3): 335-347.
}
\keyword{datasets}
