% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{treegrowth2}
\alias{treegrowth2}
\title{Crecimiento en diámetro y altura de árboles muestras de Grand-fir.}
\format{
Contiene las siguientes siete columnas:
\describe{
  \item{num.arb}{Numero identificador del árbol. Unico para cada árbol muestra.}
  \item{bosque}{Tipo forestal.}
  \item{habitat}{Clasificación de tipo de habitat.}
  \item{cod.arb}{Un codigo que combina a las siguientes columnas: num.arb-bosque-habitat}
  \item{edad}{Edad, en años.}
  \item{dap}{diámetro a la altura del pecho, en cm. Note que los decimales es debido a 
    que esta variable originalmente fue medida en pulgadas.}
  \item{htot}{Altura total, in m. Note que los decimales es debido a que esta variable
   fue originalmente medida en pies.}
 }
}
\source{
Originalmente los datos fueron cedidos por el Dr Albert Stage (Q.E.P.D) al Profesor Andrew Robinson (University
 of Idaho, USA), quien los usaba para explicar el ajuste de modelos estadisticos. El Dr Christian Salas-Eljatib
 fue un estudiante de postgrado en estadistica del Prof. Robinson en la Univ. of Idaho.
}
\usage{
data(treegrowth2)
}
\description{
Crecimiento en diámetro y altura de 66 árboles de Grand-fir. Los datos fueron derivados a partir
 de árboles muestras de analisis fustal colectados por el Dr Albert Stage (US Forest Service, Moscow, ID, USA.)
}
\examples{
head(treegrowth2)    
unique(treegrowth2$cod.arb)
}
\references{
Stage, A. R., 1963. A mathematical approach to polymorphic site index curves for 
 Grand fir. Forest Science 9(2):167–180.
}
\keyword{datasets}
