% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{treevolroble}
\alias{treevolroble}
\title{Contains tree-level variables for roble (Nothofagus obliqua) in the Rucamanque experimental forest, near Temuco, Chile.}
\format{
Contains tree-level variables, as follows:
\describe{
  \item{tree.no}{Tree number.}
\item{dbh}{Diameter at breast height, in cm.}
\item{htot}{Total height, in m.}
 \item{d6}{Upper-stem diameter at 6 m, in cm.}
\item{vtot}{Tree gross volume, in m\eqn{^{3}}{^3} with bark.}
 }
}
\source{
The data are provided courtesy of Dr Christian Salas-Eljatib (Chile).
}
\usage{
data(treevolroble)
}
\description{
These are tree-level measurement data of sample trees in the Rucamanque experimental forest, 
 near Temuco, in the Araucania region in south-centralChile, measured in 1999.  The
data are the same as in the dataframe "treevolruca", but only having observations for the 
species roble (Nothofagus obliqua).
}
\examples{
data(treevolroble)    
head(treevolroble) 
plot(vtot~dbh, data=treevolroble) 
}
\references{
- Salas C. 2002. Ajuste y validación de ecuaciones de volumen para un relicto
del bosque de Roble-Laurel-Lingue. Bosque 23(2): 81-92. \url{https://eljatib.com/publication/2002-01-01_ajuste_y_validacion_/}
}
\keyword{datasets}
