% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{treevolroble2}
\alias{treevolroble2}
\title{Variables alometricas a nivel de arbol para roble (Nothofagus obliqua) en el bosque de Rucamanque, cerca de Temuco, Chile.}
\format{
Contains tree-level variables, as follows:
\describe{
  \item{tree.no}{Tree id}
\item{dbh}{Diameter at breast height, in cm}
\item{htot}{Total height (m)}
 \item{d6}{Upper-stem diameter at 6 m, in cm}
\item{vtot}{Tree gross volume, in m^{3} with bark.}
 }
}
\source{
The data are provided courtesy of Dr Christian Salas at
the Universidad Mayor (Santiago, Chile).
}
\usage{
data(treevolroble2)
}
\description{
These are tree-level measurement data of sample trees in the Rucamanque experimental forest, 
 near Temuco, in the Araucania region in south-centralChile, measured in 1999.  The
data are the same as in the dataframe "treevolruca", but only having observations for the 
species roble (Nothofagus obliqua).
}
\examples{
data(treevolroble2)    
head(treevolroble2) 
}
\references{
Salas C. 2002. Ajuste y validacion de ecuaciones de volumen para un relicto
 del bosque de Roble-Laurel-Lingue. Bosque 23(2): 81-92.
}
\keyword{datasets}
