% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperleft.r
\name{upperleft}
\alias{upperleft}
\title{convert the first n-characters of a string to upper-case
letters.}
\usage{
upperleft(fac, n = 1)
}
\arguments{
\item{fac}{is an object of class string or factor}

\item{n}{is the number of characters to be converted of a the
string given in \code{fac}.}
}
\value{
This function returns an object having the first
n-characters from the left-hand side in upper-case.
}
\description{
Function to upper-case the first n-characters of a string from
the left-hand side.
}
\details{
It is specially set to arrange data vector having
alphanumeric (i.e., letters) format.
}
\examples{

fac.x<-"willkommen"
upperleft(fac.x)
upperleft(fac.x,n = 2)
upperleft(fac.x,2)
upperleft(fac.x,3)
#A longer vector of characters
fac.x<-c("willkommen","welcome","bem-vindo","bievenido")
upperleft(fac.x,1)
}
\references{
Salas-Eljatib, C. 2021. Análisis de datos con el programa
estadístico R: una introducción
aplicada. Ediciones Universidad Mayor, Santiago, Chile. 170 p.
\url{https://eljatib.com/rlibro}
}
\author{
Christian Salas-Eljatib
}
