% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{airquality}
\alias{airquality}
\title{Airquality data, from the datasets library.}
\format{
Contains 6 variables, as follows:
\describe{
  \item{ozone}{numeric	Ozone (ppb).}
  \item{solar}{numeric	Solar R (lang).}
\item{wind}{numeric	Wind (mph).}
\item{temp}{numeric	Temperature (degrees F).}
\item{month}{numeric	Month (1--12).}
\item{day}{numeric	Day of month (1--31).}
 }
}
\source{
The data were obtained from the library 'datasets'.
}
\usage{
data(airquality)
}
\description{
Daily air quality measurements in New York, May to September 1973.
}
\examples{
data(airquality)    
head(airquality) 
}
\references{
Chambers J, Cleveland W, Kleiner B, Tukey P. 1983. Graphical Methods for Data Analysis. Belmont. CA: Wadsworth.
}
\keyword{datasets}
