% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{annualppCities}
\alias{annualppCities}
\title{Time series of annual precipitations in cities of Chile.}
\format{
The dataframe contains three variables as follows:
\describe{
  \item{city}{Name of city.}
  \item{year}{Year of registry.}
  \item{annual}{Value of the annual precipitation of a given year (mm).}
 }
}
\source{
The data were obtained from \url{https://explorador.cr2.cl/}.
}
\usage{
data(annualppCities)
}
\description{
Data contains annual precipitations in six cities in Chile (Santiago, Talca, Chillan, Temuco, Valdivia, and Puerto Montt) at different years.
}
\examples{
data(annualppCities)    
head(annualppCities) 
}
\keyword{datasets}
