% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{annualppCities2}
\alias{annualppCities2}
\title{Serie de tiempo de precipitaciones anuales en Chile.}
\format{
The dataframe contains three variables as follows:
\describe{
  \item{ciudad}{Name of city.}
  \item{anho}{Year of registry.}
  \item{pp.anual}{Value of the annual precipitation of a given year (mm).}
 }
}
\source{
Los datos fueron obtenidos desde \url{https://explorador.cr2.cl/}.
}
\usage{
data(annualppCities2)
}
\description{
Data contains annual precipitations in six cities in Chile (Santiago, Talca, Chillan, Temuco, Valdivia, and Puerto Montt) at different years.
}
\examples{
data(annualppCities2)    
head(annualppCities2) 
}
\keyword{datasets}
