% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{araucaria2}
\alias{araucaria2}
\title{Variables a nivel de parcela para bosques de Araucaria araucana el sur de Chile.}
\format{
Contains plot-level variables as follows:
\describe{
  \item{rodal}{Stand number}
  \item{parcela}{Plot sample identificator number}
  \item{x.utm}{UTM  coordinate in X-axis, in km}
  \item{y.utm}{UTM  coordinate in Y-axis, in km}
    \item{pendiente}{Slope, in \%}
\item{exposicion}{Aspect, in degrees}
    \item{altitud}{Elevation, in msnm}    
  \item{nha}{Densidad, en arb/ha}
 \item{gha}{Area basal, en m\eqn{^{2}}{^2}/ha}
\item{hdom}{Altura dominante, en m}
\item{vha}{Volumen bruto, en m\eqn{^{3}}{^3}/ha} 
 \item{dg}{Diameter medio cuadratico, en cm}
 }
}
\source{
Los datos a nivel de árbol fueron cedidos por el Dr Nelson Ojeda de
la Universidad de La Frontera (Temuco, Chile).
}
\usage{
data(araucaria)
}
\description{
Estos son variables a nivel de parcela para bosques Araucaria araucana  
 en el sur de Chile, medidos en 2009.  Estas variables se basan en mediciones
 realizadas en parcelas de muestreo de superficie fija de 1000 m\eqn{^{2}}{^2}.  Hay dos rodales.
}
\examples{
data(araucaria2)    
head(araucaria2) 
}
\references{
Salas C, Ene L, Ojeda N, Soto H. 2010. Metodos estadísticos parametricos y no parametricos
   para predecir variables de rodal basados en Landsat ETM+: una comparacion en un bosque
    de Araucaria araucana en Chile. Bosque 31(3): 179-194.
}
\keyword{datasets}
