% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{araucaria2}
\alias{araucaria2}
\title{Variables a nivel de parcela para bosques de Araucaria araucana en Chile.}
\format{
Contiene las siguientes variables:
\describe{
  \item{rodal}{Rodal, con un número indentificador.}
  \item{parce}{Parcela de muestreo, con un número indentificador.}
  \item{x.utm}{Coordenada UTM en el eje X, en km.}
  \item{y.utm}{Coordenada UTM en el eje Y, en km.}
    \item{pendiente}{Pendiente, en \%}
\item{exposicion}{Exposición del terreno, en grados.}
    \item{altitud}{Altitud, en msnm.}    
  \item{nha}{Densidad, en arb/ha.}
 \item{gha}{Área basal, en m\eqn{^{2}}{^2}/ha.}
\item{hdom}{Altura dominante, en m.}
\item{vha}{Volumen bruto, en m\eqn{^{3}}{^3}/ha.} 
 \item{dg}{Diámetro del árbol de área basal media, en cm.}
 }
}
\source{
Los datos a nivel de árbol fueron cedidos por el Dr. Nelson Ojeda de
la Universidad de La Frontera (Temuco, Chile).
}
\usage{
data(araucaria)
}
\description{
Estos son variables a nivel de parcela para bosques Araucaria araucana  
 en el centro-sur de Chile, medidos en 2009.  Estas variables se basan 
 en mediciones
 realizadas en parcelas de muestreo de 1000 m\eqn{^{2}}{^2}. Hay dos rodales.
}
\examples{
data(araucaria2)    
head(araucaria2) 
}
\references{
Salas C, Ene L, Ojeda N, Soto H. 2010. Metodos estadísticos 
parametricos y no parametricos
   para predecir variables de rodal basados en Landsat ETM+: 
   una comparacion en un bosque
    de Araucaria araucana en Chile. Bosque 31(3): 179-194.
     \doi{10.4067/S0717-92002010000300002}
}
\keyword{datasets}
