% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{biomass}
\alias{biomass}
\title{Contains tree-level biomass data for several species in Canada.}
\format{
\describe{
  \item{treenum}{tree number.}
  \item{spp}{species common name.}
  \item{dbh}{diameter at breast height, in cm.}
  \item{height}{total height, in m.}
  \item{totbiom}{total biomass, in kg.}
  \item{bolebiom}{stem biomass, in kg.}
  \item{branchbiom}{branches biomass, in kg.}
  \item{foliagebiom}{foliage biomass, in kg.}
 }
}
\source{
El archivo de datos fue preparado, y referido, por el Prof. Timothy 
Gregoire de
 Yale University (New Haven, CT, USA), mientras Christian Salas
  (el autor del presente paquete fue su Teaching Assistant).
}
\usage{
biomass
}
\description{
These are tree-level variables for several species in Canada.
}
\examples{
data(biomass)    
head(biomass)
tapply(biomass$totbiom,biomass$spp,summary)  
}
\keyword{datasets}
