% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{election2}
\alias{election2}
\title{Elección presidencial en el estado de Florida (USA) en el 2000.}
\format{
Contiene las siguientes tres columnas:
\describe{
  \item{gore}{Votos para Gore. Número de votos para Al Gore.}
  \item{bush}{Votos para Bush. Número de votos para George W. Bush.}
\item{buchanan}{Votos para Buchaman. Número de votos para Pat Buchanan.}
 }
}
\source{
Los datos se obtuvieron desde el paquete 'alr4' de R.
}
\usage{
data(election2)
}
\description{
Conteo de votos a nivel de condado en el estado de Florida, año 2000.
}
\examples{
data(election2)    
head(election2) 
}
\references{
Weisberg S. 2014. Applied Linear Regression. 4th edition. Hoboken NJ: Wiley
}
\keyword{datasets}
