% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{election2}
\alias{election2}
\title{Eleccion presidential en el estado de Florida (USA) en el 2000.}
\format{
Contains 3 variables, as follows:
\describe{
  \item{gore}{Votos para Gore.}
  \item{bush}{Votos para Bush.}
\item{buchaman}{Votos para Buchaman.}
 }
}
\source{
Los datos se obtuvieron desde el paquete 'alr4' de R.
}
\usage{
data(election2)
}
\description{
County-by-county vote for president in Florida in 2000 for Bush, Gore and Buchanan.
}
\examples{
data(election2)    
head(election2) 
}
\references{
Weisberg S. 2014. Applied Linear Regression. 4th edition. Hoboken NJ: Wiley
}
\keyword{datasets}
