% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{eucaleaf2}
\alias{eucaleaf2}
\title{Mediciones foliares para arboles de Eucalyptus nitens en Tasmania, Australia.}
\format{
Contiene variables a nivel de hoja, como sigue:
\describe{
  \item{tiempo}{factor a dos niveles: Temprano o Tardio}
\item{arbol}{un identificador del arbol muestra}
\item{meristema}{descripcion del meristema}
\item{largo}{largo de la hoja, en mm}
\item{ancho}{ancho de la hoja, en mm}
\item{area}{area foliar, en cm^2}
 }
}
\source{
Aunque la fuente original de estas mediciones proviene de la tesis del Dr Candy (1999), 
 el archivo de datos fue cortesia del Prof. Timothy Gregoire de
 Yale University (New Haven, CT, USA). Ademas, estos datos fueron ocupados en
 el estudio de Gregoire y Salas (2009).
}
\usage{
data(eucaleaf)
}
\description{
Mediciones de largo, ancho y area de hojas de Eucalyptus nitens.
}
\examples{
data(eucaleaf2)    
head(eucaleaf2) 
}
\references{
- Candy SG. 1999. Predictive models for integrated pest management of the
leaf beetle Chrysophtharta bimaculata in Eucalyptus nitens in Tasmania. Doctoral dissertation,
 University of Tasmania, Hobart, Australia.

- Gregoire TG, y Salas C. 2009. Ratio estimation with measurement
error in the auxiliary variate. Biometrics 65(2):590-598
}
\keyword{datasets}
