% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{eucaplot}
\alias{eucaplot}
\title{Data from a Eucalyptus globulus plantation near Gorbea, Region de La Araucania, Chile.}
\format{
The dataframe contains four variables as follows:
\describe{
  \item{dbh}{Diameter at breast height, in cm.}
  \item{health}{health status (1: good, 2: medium, 3: bad).}
  \item{shape}{stem shape for timber purposes (1: good, 2: medium, 3: bad).}
  \item{crown.class}{Crown class (1: superior, 2: intermedium, 3: lower).}
  \item{toth}{Total height, in m.}
 }
}
\source{
The data were provided courtesy of Dr Christian Salas-Eljatib (Chile).
}
\usage{
data(eucaplot)
}
\description{
Tree-level data collected within a sample plot in a forestry plantation of Eucalyptus globulus near Gorbea, 
Southern Chile. The plot size is 500 square meters. The plantation is 15 yr-old and
had been subject to three thinnings.
}
\examples{
data(eucaplot)    
head(eucaplot) 
}
\keyword{datasets}
