% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{eucaplot2}
\alias{eucaplot2}
\title{Árboles dentro de parcelas de muestreo en una plantación de Eucalyptus globulus, Chile.}
\format{
Los datos contienen las siguientes cuatro columnas:
\describe{
  \item{dap}{Diámetro a la altura del pecho, en cm.}
  \item{sanidad}{Evaluación cualitativa de la sanidad del árbol (1: buena, 2: media, 3: mala).}
  \item{forma}{Evaluación cualtitativa de la forma del fuste (1: buena, 2: media, 3: mala).}
  \item{clase.copa}{Clase de copa (1: superior, 2: intermedio, 3: inferior).}
  \item{atot}{Altura total, en m.}
 }
}
\source{
Los datos fueron cedidos por el Prof. Christian Salas (Universidad de Chile, Santiago, Chile).
}
\usage{
data(eucaplot2)
}
\description{
Datos a nivel de árbol colectados dentro de una parcela de muestreo en una plantación
  forestal de Eucalyptus globulus cerca de Gorbea, Sur de Chile. La superficie de la parcela es de 500 m\eqn{^{2}}{^2}.
  La plantación tiene 15 años de edad y ha estado sujeta a tres raleos.
}
\examples{
data(eucaplot2)    
head(eucaplot2) 
}
\keyword{datasets}
