% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRight.r
\name{extractRight}
\alias{extractRight}
\title{Extracts the last n-characters of a string}
\usage{
extractRight(fac, n)
}
\arguments{
\item{fac}{is an object of class string or factor}

\item{n}{is the number of characters to be deleted of a the
string given in 'fac'.}
}
\value{
This function returns an object having the last n
characters from the right-hand side.
}
\description{
Function to extract the last n-characters of a string from the
right-hand side.
}
\details{
It is specially set to arrange data vector
having alphanumeric format.
}
\examples{

last.names.id <- c("Stage-1924","Gregoire-1958","Robinson-1967")
extractRight(last.names.id,4)
extractRight(last.names.id,2)
}
\references{
Salas-Eljatib, C. 2021. Análisis de datos con el programa
estadístico R: una introducción
aplicada. Ediciones Universidad Mayor, Santiago, Chile. 170 p.
\url{https://eljatib.com/rlibro}
}
\author{
Christian Salas-Eljatib
}
