% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{fishgrowth2}
\alias{fishgrowth2}
\title{Crecimiento de peces}
\format{
Contains 3 variables, as follows:
\describe{
  \item{edad}{Year at capture.}
  \item{largo}{Length at capture (mm).}
\item{escala}{radius of a key scale (mm).}
 }
}
\source{
Datos obtenidos desde el paquete alr4 de R.
}
\usage{
data(fishgrowth2)
}
\description{
Data on samples of small mouth bass collected in West Bearskin Lake, Minnesota, in 1991. The
file wblake includes only fish of ages 8 or younger.
}
\examples{
data(fishgrowth2)    
head(fishgrowth2) 
}
\references{
Weisberg S. 2014. Applied Linear Regression. 4th edition. Hoboken NJ: Wiley
}
\keyword{datasets}
