% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{football}
\alias{football}
\title{Anaerobic potential of soccer players.}
\format{
The data frame contains 13 variables as follows:
\describe{
  \item{WPM}{}											
\item{WPMk}{}
\item{WPm}{}
\item{WPmk}{}
\item{WTT}{}
\item{WTTk}{}
\item{WIF}{}
\item{W5}{}
\item{W10}{}
\item{W15}{}
\item{W20}{}
\item{W25}{}
\item{W30}{}
 }
}
\source{
Data were provided by Dr Aquiles Yanez-Silva at Universidad
 Mayor (Santiago, Chile).
}
\usage{
data(football)
}
\description{
Data about anaerobic variables of football players.
}
\examples{
data(football)    
head(football) 
}
\references{
Not yet.
}
\keyword{datasets}
