% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{football2}
\alias{football2}
\title{Potencia anaerobica de jugadores de football.}
\format{
Contiene variables de nivel individual, como se describen a continuacion::
\describe{
  \item{WPM}{Watt Peak Max.}											
\item{WPMk}{Watt Peak Max relative.}
\item{WPm}{Watt Peak media.}
\item{WPmk}{Watt Peak media relative.}
\item{WTT}{Watt Total Work.}
\item{WTTk}{Watt Total Work relative.}
\item{WIF}{Watt Index Fatigue.}
\item{W5}{Watt 5 seg.}
\item{W10}{Watt 10 seg.}
\item{W15}{Watt 15 seg.}
\item{W20}{Watt 20 seg.}
\item{W25}{Watt 25 seg.}
\item{W30}{Watt 30 seg.}
 }
}
\source{
Los datos fueron cedidos por el Dr Aquiles Yanez-Silva de la Universidad Mayor (Santiago, Chile).
}
\usage{
data(football2)
}
\description{
Datos sobre potencia anaerobica de jugadores de football.
}
\examples{
data(football2)    
head(football2) 
}
\references{
Not yet.
}
\keyword{datasets}
