% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{idahohd}
\alias{idahohd}
\title{Tree height-diameter data from Idaho (USA).}
\format{
Contains five variables, as follows:
\describe{
  \item{plot}{plot number.}
  \item{tree}{tree within plot.}
\item{species}{a factor with levels DF = Douglas-fir, GF = Grand fir, SF = Subalpine fir, WL = Western larch,
                WC = Western red cedar, WP = White pine.}
\item{dbh}{Diameter 137 cm perpendicular to the bole, mm.}
\item{height}{Height of the tree, in decimeters.}
 }
}
\source{
The data were obtained from the alr4 library.
}
\usage{
data(idahohd)
}
\description{
These data are forest inventory measures from the Upper Flat Creek stand of the University 
of Idaho Experimental Forest, dated 1991.
}
\examples{
data(idahohd)    
head(idahohd) 
}
\references{
Weisberg S. 2014. Applied Linear Regression. 4th edition. New York: Wiley.
}
\keyword{datasets}
