% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{idahohd2}
\alias{idahohd2}
\title{Altura-diametro de arboles en el estado de Idaho (USA).}
\format{
Contiene cinco variables, como sigue:
\describe{
  \item{parce}{numero de la parcela de muestreo.}
  \item{narb}{numero del arbol dentro de la parcela.}
\item{spp}{especie del arbol, una variable factor  con niveles DF = Douglas-fir, GF = Grand fir, SF = Subalpine fir, WL = Western larch,
                WC = Western red cedar, WP = White pine.}
\item{d}{Diametro del tronco a los 1.3m sobre el suelo, en cm.}
\item{h}{Altura del arbol, en m.}
 }
}
\source{
Los datos fueron obtenidos desde la libreria alr4.
}
\usage{
data(idahohd2)
}
\description{
Estos datos provienen de un muestreo en un bosque de
Upper Flat Creek en el bosque experimental de la 
University of Idaho, medido en 1991.
}
\examples{
data(idahohd2)    
head(idahohd2) 
}
\references{
Weisberg S. 2014. Applied Linear Regression. 4th edition. New York: Wiley.
}
\keyword{datasets}
