% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDatas.R
\docType{data}
\name{lleuque}
\alias{lleuque}
\title{Contains tree density by species and plot for Prumnopitys andina (Lleuque) 
forests}
\format{
The dataframe has the following columns
\describe{
  \item{stand}{Stand number}
  \item{plot.num}{Sample plot number}
  \item{cipres}{Tree density of ciprés de la Cordillera
   (Austrocedrus chilensis), in trees/ha.}
  \item{ldura}{Tree density of leña dura (Maytenus disticha), in trees/ha.}
  \item{roble}{Tree density of roble (Nothofagus obliqua), in trees/ha.}
  \item{lleuque}{Tree density of lleuque (Prumnopitys andina), in trees/ha.}
 }
}
\source{
The data are provided courtesy of Prof. Rodrigo Vargas-Gaete at Universidad
 de La Frontera (Temuco, Chile).
}
\usage{
lleuque
}
\description{
Contains species composition data for forests with presence of Lleuque 
(Prumnopitys andina)
}
\references{
Vargas-Gaete R, Salas-Eljatib C, Penneckamp D, Neira Z, Diez MC,
 Vargas-Picón, R. 2020. Estructura 
y regeneración de bosques de Prumnopitys andina en los Andes del sur
 de Chile. Gayana Botánica
 77(1):48-58. \doi{10.4067/S0717-66432020000100048}
}
\keyword{datasets}
