% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{moda}
\alias{moda}
\title{Computes the mode}
\usage{
moda(y = y)
}
\arguments{
\item{y}{is a numeric vector.}
}
\value{
The function returns the mode, a numeric scalar.
}
\description{
Computes the mode of a random variable.
}
\details{
The mode is an statistics representing the most "used" value of
the random variable as a way of central position.
}
\examples{

set.seed(1234)
variable <- rnorm(10, mean=45,sd=6)
#using the function
moda(y=variable)
moda(variable)
}
\author{
Christian Salas-Eljatib.
}
