% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{pinaster}
\alias{pinaster}
\title{Tree volume for Pinus pinaster in the Baixo-Miño, Galicia, Spain.}
\format{
Contains tree-level variables, as follows:
\describe{
  \item{stand}{Stand from where the sample tree was collected.}
\item{si}{Site index of the stand, in m.}
\item{tree.no}{Tree number.}
\item{dbh}{Diameter at breast height, in cm.}
\item{htot}{Total height, in m.}
 \item{d4}{Upper-stem diameter at 4 m, in cm.}
\item{vol.wb}{Tree gross volume, in m\eqn{^{3}}{^3} with bark.}
\item{vol.wob}{Tree gross volume, in m\eqn{^{3}}{^3} without bark.}
 }
}
\source{
The data are provided courtesy of Dr Christian Salas-Eljatib.
}
\usage{
data(pinaster)
}
\description{
Tree-level volume data of sample trees in the Baixo-Miño region, Galicia, Spain.
}
\examples{
data(pinaster)    
head(pinaster) 
}
\references{
- Salas C, Nieto L, Irisarri A. 2005. Modelos de volumen para Pinus pinaster Ait. en la comarca del Baixo Mino,
  Galicia, Espana. Quebracho 12: 11-22. \url{https://eljatib.com/publication/2005-12-01_modelos_de_volumen_p/}
}
\keyword{datasets}
