% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{presidentChile}
\alias{presidentChile}
\title{2021 presidential election in Chile.}
\format{
The dataframe contains the following columns:
\describe{
  \item{region.no}{Administrative region number of Chile.}
  \item{region}{Administrative region name.}
  \item{province}{Province.}
  \item{senat.const}{Senatorial constituency.}
  \item{distrit}{Distrit.}
  \item{county}{County.}
  \item{elect.const}{Electoral constituency.}
  \item{location}{Location.}
  \item{table}{Voter table.}
  \item{table.type}{Voter table type.}
  \item{merged.tab}{Merged voter tables.}
  \item{electors}{Electors.}
  \item{num.in.vote}{.}
  \item{candidate}{Candidate. Gabriel Boric and Jose A. Kast}
  \item{total.votes}{Total number of votes acording to the TRICEL.}       
 }
}
\source{
The data were obtained from the electoral service of the Chilean Government (SERVEL) at
 \url{https://www.servel.cl}.
  The datafile name was "Resultados_mesa_presidencial_TRICEL_2v_2021-1.xlsx", and was downloaded
  on October 24, 2022.
}
\usage{
data(presidentChile)
}
\description{
Voter table-level data of the 2021 presidential election in Chile. The election was held on December 19, 2021.
}
\examples{
data(presidentChile)    
head(presidentChile) 
}
\keyword{datasets}
