% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{presidentChile}
\alias{presidentChile}
\title{2021 presidential election in Chile.}
\format{
The data frame contains the following columns:
\describe{
  \item{region.no}{Administrative region number of Chile.}
  \item{region}{Administrative region name.}
  \item{province}{Province.}
  \item{circu.senate}{Senatorial constituency.}
  \item{distrite}{Distrit.}
  \item{county}{County.}
  \item{elec.distri}{Electoral constituency.}
  \item{local}{Place.}
  \item{table.no}{Voting table number.}
  \item{table.type}{Voting table type: M (women), V (men).}
  \item{merged.tables}{Merged voting tables.}
  \item{electors}{Electors.}
  \item{number.in.voto}{Numero del candidato en el voto.}
  \item{candidate}{Candidate: Gabriel Boric, Jose A. Kast, Nule votes, Empty votes.}
  \item{total.votes}{Total number of votes.}       
 }
}
\source{
The data were obtained from the electoral service of the Chilean Government (SERVEL) at
 \url{https://www.servel.cl/centro-de-datos/procesos-electorales-detalle/?id=27}.
  The datafile name was "Resultados_mesa_presidencial_TRICEL_2v_2021-1.xlsx", and was downloaded
  on October 24, 2022.
}
\usage{
data(presidentChile)
}
\description{
Voting table-level data of the 2021 presidential election in Chile. The election was held on December 19, 2021.
}
\examples{
data(presidentChile)    
head(presidentChile) 
}
\keyword{datasets}
