% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{pspRuca2}
\alias{pspRuca2}
\title{Medidas a nivel de arbol y coordenadas espaciales en predio rucamanque}
\format{
Variables se describen a continuacion::
\describe{
\item{tree.no}{Numero de arbol}											
\item{spp}{Nombre de la especie, "N. obliqua" es Nothofagus obliqua, "Ap" es Aexitocicum puncatatum, etc.}
\item{status}{1 vivo, 0 muerto en pie}
\item{dbh}{Diámetro a la altura del pecho, en cm}
\item{x.coord}{Posicion cartesiana en el eje X, en m}
\item{y.coord}{Posicion cartesiana en el eje Y, en m}
\item{crown.class}{Clase de corona (1: superior, 2: intermedio, 3; inferior)}
 }
}
\source{
Los datos fueron cedidos por el Dr Christian Salas-Eljatib (Santiago, Chile).
}
\usage{
data(pspRuca2)
}
\description{
Medidas a nivel de arbol y coordenadas espaciales en un parcela de muestreo permanente de 1 ha (100 x 100m) en el Rucamanque 
Bosques experimentales, cerca de Temuco, Chile.
}
\examples{
data(pspRuca2)    
head(pspRuca2) 
}
\references{
Salas C, LeMay V, Nunez P, Pacheco P, and Espinosa A. 2006. Spatial patterns in an old-growth Nothofagus obliqua forest in 
south-central Chile. Forest Ecology and Management 231(1-3): 38-46.
}
\keyword{datasets}
