% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_datas.R
\docType{data}
\name{treevolruca2}
\alias{treevolruca2}
\title{Variables alometricas a nivel de arbol para varias especies del bosque de Rucamanque, cerca de Temuco, Chile.}
\format{
Las siguientes columnas son parte de la dataframe:
\describe{
  \item{arbol}{Numero del arbol.}
\item{especie}{Especie.}
\item{dap}{Diámetro a la altura del pecho, en cm.}
\item{htot}{Altura total, en m.}
 \item{d6}{Diámetro fustal a los 6 m, en cm.}
\item{vtot}{Volumen bruto total, en m\eqn{^{3}}{^3} with bark.}
 }
}
\source{
Los datos fueron cedidos por el Dr Christian Salas-Eljatib (Chile).
}
\usage{
data(treevolruca2)
}
\description{
Volumen, altura y diámetro, entre otras para árboles muestra en el bosque de Rucamanque, cerca de Temuco, 
en la region de la Araucania, en el sur de Chile.  Las siguientes especies son parte de los datos: laurel (laurelia sempervirens), lingue (Persea lingue), olivillo (Aextocicon puncatum), 
 roble (Nothofagus obliqua), tepa (Laureliosis philippiana), y tineo (Weinmannia trichosperma).
}
\examples{
data(treevolruca2)    
head(treevolruca2) 
}
\references{
- Salas C. 2002. Ajuste y validación de ecuaciones de volumen para un relicto
del bosque de Roble-Laurel-Lingue. Bosque 23(2): 81-92. \doi{10.4067/S0717-92002002000200009} \url{https://eljatib.com/publication/2002-07-01_ajuste_y_validacion_/}
}
\keyword{datasets}
