% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyBoxplot.R
\name{xyBoxplot}
\alias{xyBoxplot}
\title{xyBoxplot: Function for a scatter-plot with boxplots}
\usage{
xyBoxplot(x = x, y = y, col.dots = "blue", xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{A numeric vector representing the X-axis variable}

\item{y}{A numeric vector representing the Y-axis variable}

\item{col.dots}{(optional) A string specifying the dot colors. Default is "blue".}

\item{xlab}{(optional) A string specifying X-axis label.}

\item{ylab}{(optional) A string specifying Y-axis label.}
}
\value{
Result of calculation
}
\description{
Creates a scatter-plot with boxplots for the Y-axis variable.
}
\examples{
library(datana)
data(fishgrowth)
df <- fishgrowth
xyBoxplot(x=df$age,y=df$length)
xyBoxplot(x=df$age,y=df$length)
}
